% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{D_ex1}
\alias{D_ex1}
\title{A toy SE object with duplicated samples}
\format{
A \code{SummarizedExperiment} object with the following fields:
\describe{
  \item{assay}{Dataframe with 100 rows and 100 columns}
  \item{colData}{Dataframe with 100 rows and 2 columns}
  \item{rowData}{Dataframe with 100 rows and 0 columns}
  \item{metadata}{empty list}
}
}
\usage{
D_ex1
}
\description{
A \code{SummarizedExperiment} object with 100 samples and 100 metabolites. 5 out of the 100 samples are duplicated. Created with randomly
generated data.
}
\keyword{datasets}
