% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{D_ex2}
\alias{D_ex2}
\title{A toy SE object with duplicated metabolites}
\format{
A \code{SummarizedExperiment} object with the following fields:
\describe{
  \item{assay}{Dataframe with 50 rows and 50 columns}
  \item{colData}{Dataframe with 50 rows and 6 columns}
  \item{rowData}{Dataframe with 50 rows and 15 columns}
  \item{metadata}{empty list}
}
}
\usage{
D_ex2
}
\description{
A \code{SummarizedExperiment} object with 50 samples and 50 metabolites. 5 out of the 100 samples are duplicated. The data for this object
is subset from the simulated_data.xlsx file.
}
\keyword{datasets}
