% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_apply.R
\name{mt_anno_apply}
\alias{mt_anno_apply}
\title{Apply function to feature or sample annotation column}
\usage{
mt_anno_apply(D, anno_type, col_name, fun)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{anno_type}{Either "samples" (colData) or "features" (rowData).}

\item{col_name}{Name of column to apply function to.}

\item{fun}{Function to be applied.}
}
\value{
colData or rowData: Changes the contents of a column.
}
\description{
Works in the same way as lapply/sapply, with a function to be executed on each entry of a given feature or sample annotation.
}
\examples{
 \dontrun{... \%>\%
 # ensure factor for casecontrol variable
 mt_anno_apply(anno_type='samples',
               col_name='casecontrol',
               fun=as.factor) \%>\%
 ...}

}
\author{
JK
}
