% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_hmdb_to_kegg.R
\name{mt_anno_hmdb_to_kegg}
\alias{mt_anno_hmdb_to_kegg}
\title{Create KEGG identifiers from HMDB identifiers}
\usage{
mt_anno_hmdb_to_kegg(D, in_col, out_col = "KEGG")
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{in_col}{Name of the rowData column containing the HMDB identifiers.}

\item{out_col}{Name of the new rowData column to be created with the KEGG identifiers. Default="KEGG".}
}
\value{
rowData: New annotation column added.
}
\description{
Creates annotations and merges them into current \code{SummarizedExperiment} object. Performs "left-joins", i.e. leaves the
original SE unchanged and just adds information where it can be mapped.
}
\examples{
# Load data, sheet with sample annotations, sheet with metabolite annotations and add KEGG identifiers
\dontrun{D <-
  # load raw data
  mt_load_xls(file=file, sheet="data", samples_in_rows=T, id_col="SAMPLE_NAME") \%>\%
  # sample annotations from metabolomics run
  mt_anno_xls(file=file, sheet="sampleinfo", anno_type="samples", anno_id_col = "SAMPLE_NAME") \%>\%
  # metabolite annotations
  mt_anno_xls(file=file, sheet="metinfo", anno_type="metabolites", anno_id_col="BIOCHEMICAL", data_id_col = "name") \%>\%
  # add KEGG identifiers
  mt_anno_hmdb_to_kegg(in_col="HMDb_id", out_col="KEGG") \%>\%
  ...}

}
\author{
EB
}
