% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_missingness.R
\name{mt_anno_missingness}
\alias{mt_anno_missingness}
\title{Add missingness annotation column}
\usage{
mt_anno_missingness(D, anno_type, out_col = "missingness")
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{anno_type}{Either "samples" (colData) or "features" (rowData).}

\item{out_col}{Name of new missingness column. Default: "missingness".}
}
\value{
rowData or colData: New annotation column added.
}
\description{
Adds a rowData or colData column representing the missingness of features or samples, respectively.
}
\examples{
\dontrun{
   # load data
   mt_load_metabolon(file=file, sheet="data") \%>\%
   # add feature missingness column to rowData
   mt_anno_missingness(anno_type = "features", out_col = "Feat_Missing")
   # add sample missingness column to colData
   mt_anno_missingness(anno_type = "samples", out_col = "Samp_Missing")
}

}
\author{
KC
}
