% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_mutate.R
\name{mt_anno_mutate}
\alias{mt_anno_mutate}
\title{Create new variable by expression}
\usage{
mt_anno_mutate(D, anno_type, col_name, term)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{anno_type}{Either "samples" (colData) or "features" (rowData).}

\item{col_name}{Name of new (or existing) annotation column to store result of expression.}

\item{term}{Mutate term to forward to \code{dplyr::mutate()}. Can include any columns from colData or rowData.}
}
\value{
colData or rowData: New (or existing) annotation column added (or replaced).
}
\description{
Creates a new variable using dplyr's mutate() mechanism. Formula can use any column in the respective colData/rowData. Can create
a new annotation column or replace an existing one (if col_name and term arguments refer to same column).
}
\examples{
\dontrun{# Convert numeric sample annotation field 'num' to factor
...  \%>\%
 mt_anno_mutate(anno_type="samples", col_name="num", term = as.factor(num)) \%>\% ...}

}
\author{
JK
}
