% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_pathways_graphite.R
\name{mt_anno_pathways_graphite}
\alias{mt_anno_pathways_graphite}
\title{Add pathway information}
\usage{
mt_anno_pathways_graphite(
  D,
  in_col,
  out_col,
  pwdb_species,
  pwdb_name,
  n_cores = 2,
  raw_db_outfile
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{in_col}{The rowData column to use for pathway fetching. The selected column must contain metabolite
identifiers (e.g. HMBD, KEGG, ChEBI, etc).}

\item{out_col}{New column name for rowData to output pathway information to.}

\item{pwdb_species}{Name of the species the data was measured in. Use \code{pathwayDatabases()} after loading graphite to
see a full list of databases and species.}

\item{pwdb_name}{Name of the pathway database to use. Use \code{pathwayDatabases()} after loading graphite to see a full
list of databases and species.}

\item{n_cores}{Number of cores to use for parallelizaion. Used in \code{convertIdentifiers()}. Default: 2.}

\item{raw_db_outfile}{OPTIONAL. Name of file to export the pathway database to. Must be a string containing the path name with a
.xlsx extension.}
}
\value{
rowData: New pathway annotation column added.

$results$pathways: A dataframe of pathway information.
}
\description{
Adds custom pathways to the already existing \code{SummarizedExperiment} data structure using the \code{graphite} package.
}
\examples{
# annotate metabolites using the humancyc database from the graphite package
\dontrun{... \%>\%
    mt_anno_pathways_graphite(in_col = "KEGG",
                              out_col = "humancyc_db",
                              pwdb_species = "hsapiens",
                              pwdb_name = "humancyc",
                              n_cores = 5) \%>\%
...}

}
\author{
PG
}
