% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_pathways_hmdb.R
\name{mt_anno_pathways_hmdb}
\alias{mt_anno_pathways_hmdb}
\title{Add pathway information}
\usage{
mt_anno_pathways_hmdb(
  D,
  in_col,
  out_col,
  pwdb_name = "SMP",
  db_dir = system.file("extdata", "precalc/hmdb/", package = "maplet"),
  db_file,
  raw_db_outfile
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{in_col}{The rowData column to use for pathway fetching. The selected column must contain HMDB metabolite
identifiers.}

\item{out_col}{Name for rowData to output pathway information to.}

\item{pwdb_name}{Name of the pathway database to use. Can use either SMP or KEGG for this argument. Default: "SMP".}

\item{db_dir}{Name of the directory where the parsed HMDB files are stored.}

\item{db_file}{File name of the parsed HMDB file to use. Must be in the format: hmdb_preprocessed_{version_number}.rds.}

\item{raw_db_outfile}{OPTIONAL. Name of file to export the pathway database to. Must be a string containing the path name with a
.xlsx extension.}
}
\value{
rowData: New pathway annotation column added.

$results$pathways: A dataframe of pathway information.
}
\description{
Adds custom pathways to the already existing \code{SummarizedExperiment} data structure using the HMDB .xml file.
}
\examples{
# annotate metabolites using smp_db
\dontrun{... \%>\%
  mt_anno_pathways_hmdb(in_col = "HMDb_ID",
                        out_col = "smp_db",
                        pwdb_name = "SMP",
                        db_dir = system.file("extdata", "precalc/hmdb/", package = "maplet")) \%>\%
...}

}
\author{
PG
}
