% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_pathways_remove_redundant.R
\name{mt_anno_pathways_remove_redundant}
\alias{mt_anno_pathways_remove_redundant}
\title{Remove redundant pathway annotations}
\usage{
mt_anno_pathways_remove_redundant(D, feat_col, pw_col)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{feat_col}{Column containing feature IDs.}

\item{pw_col}{Column containing pathway IDs.}
}
\value{
rowData: Redundant pathway annotation from SE pw_col column filtered.
}
\description{
Remove identical pathways from an existing \code{SummarizedExperiment} data structure that has a column of pathway annotations.
}
\examples{
# first annotate features using smp_db and then remove redundant pathways
\dontrun{... \%>\%
  mt_anno_pathways_HMDB(in_col = "HMDb_ID",
                         out_col = "smp_db",
                         pwdb_name = "SMP",
                         db_dir = system.file("extdata", "precalc/hmdb/", package = "maplet")) \%>\%
  mt_anno_pathways_remove_redundant(feat_col = "HMDb_ID", pw_col = "smp_db") \%>\%
...}

}
\author{
PG
}
