% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_pathways_uniprot.R
\name{mt_anno_pathways_uniprot}
\alias{mt_anno_pathways_uniprot}
\title{Add pathway information}
\usage{
mt_anno_pathways_uniprot(D, in_col, out_col, raw_db_outfile)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{in_col}{Column to use for pathway fetching. The selected column must contain protein Uniprot identifiers.}

\item{out_col}{A new column name for rowData to output pathway information to.}

\item{raw_db_outfile}{OPTIONAL. Name of file to export the pathway database to. Must be a string containing the path name with a
.xlsx extension.}
}
\value{
rowData: New pathway annotation column for proteins.

$results$pathways: A dataframe of pathway information.
}
\description{
Modified from mt_anno_pathways_hmdb. Adds custom pathways to the already existing \code{SummarizedExperiment} data taken from
KEGGREST.
}
\examples{
# annotate proteins using kegg
\dontrun{... \%>\%
  mt_anno_pathways_uniprot(in_col = "Uniprot_ID", out_col = "kegg_db") \%>\%
...}

}
\author{
RB
}
