% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_pathways_xls.R
\name{mt_anno_pathways_xls}
\alias{mt_anno_pathways_xls}
\title{Add pathway information}
\usage{
mt_anno_pathways_xls(
  D,
  file,
  in_col,
  out_col,
  sheet,
  feat_id_col,
  pw_id_col,
  pw_name_col,
  raw_db_outfile
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{file}{Path where the pathway annotation flat file is stored.}

\item{in_col}{rowData column to use for pathway fetching. The selected column must contain feature
identifiers (e.g. HMBD, KEGG, ChEBI, etc).}

\item{out_col}{New column name for rowData to output pathway information to.}

\item{sheet}{Sheet name or number to read in flat file.}

\item{feat_id_col}{Name of flat file column containing feature IDs.}

\item{pw_id_col}{Name of flat file colname containing pathway IDs.}

\item{pw_name_col}{Name of flat file colname containing pathway names.}

\item{raw_db_outfile}{OPTIONAL. Name of file to export the pathway database to. Must be a string containing the path name with a
.xlsx extension.}
}
\value{
rowData: New pathway annotation column added.

$results$pathways: A dataframe of pathway information.
}
\description{
Adds custom pathways to the already existing \code{SummarizedExperiment} data structure using a flat file.
NOTE: The flat file should be an Excel file, containing 3 columns:
\itemize{
  \item{feat_id_col} contains feature IDs
  \item{pw_id_col} contains pathway IDs
  \item{pw_name_col} contains pathway names}
The pathway columns pw_id_col and pw_name_col should be in long format, meaning if a feature is assigned to multiple pathways,
each feature pathway pair should appear in a separate row.
}
\examples{
\dontrun{# annotate features using the SMP column of the pathway database flat file
... \%>\%
      mt_anno_pathways_xls(file = system.file("extdata", "precalc/hmdb/hmdb_preprocessed_4.0.xlsx", package = "maplet"),
                           in_col = "HMDb_ID",
                           out_col = "janpw",
                           feat_id_col = "HMDB_id",
                           pw_id_col = "SMP",
                           pw_name_col = "pathway_name") \%>\%
...}

}
\author{
PG
}
