% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_rename.R
\name{mt_anno_rename}
\alias{mt_anno_rename}
\title{Rename selected columns}
\usage{
mt_anno_rename(D, anno_type, old_col_names, new_col_names)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{anno_type}{Either "samples" (colData) or "features" (rowData).}

\item{old_col_names}{Vector of names of existing annotation columns to be changed.}

\item{new_col_names}{Vector of new names to replace existing names.}
}
\value{
colData or rowData: Column names replaced.
}
\description{
Replaces existing column names with new column names using dplyr's rename() function. Both old
and new column names should be provided as a vector. Works on either colData or rowData
annotation data frames.
}
\examples{
\dontrun{# Replace column names in colData data frame
...  \%>\%
 mt_anno_rename(anno_type="samples", old_col_names=c(), new_col_names = c()) \%>\% ...}

}
\author{
KC
}
