% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_reorder_factor.R
\name{mt_anno_reorder_factor}
\alias{mt_anno_reorder_factor}
\title{Reorder a sample annotation}
\usage{
mt_anno_reorder_factor(D, col_name, new_order)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{col_name}{Name of the column in colData to access.}

\item{new_order}{List of strings containing the new order.}
}
\value{
colData: Changes the factor order of one column.
}
\description{
Does not change anything in the actual dataset, just the factor order of a column in colData. Can be used to change the plotting
order in boxplots, volcano plots, etc.
}
\examples{
\dontrun{\%>\% mt_anno_reorder_factor(col_name="Group",
                                    new_order=c('WT_Norm_F','WT_Hyp_F','KO_Norm_F','KO_Hyp_F','WT:KO2_F','WT:WT2_F')) \%>\%
}

}
\author{
JK
}
