% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_select.R
\name{mt_anno_select}
\alias{mt_anno_select}
\title{Select columns from annotation data frame}
\usage{
mt_anno_select(D, anno_type, term)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{anno_type}{Either "samples" (colData) or "features" (rowData).}

\item{term}{Selection expression to forward to \code{dplyr::select()}. Can include any columns
from colData or rowData.}
}
\value{
colData or rowData: Column(s) selected or removed.
}
\description{
Selects a subset of columns from either the colData or rowData annotation data frame. The user
can either choose to provide a list of columns to keep or to provide a list of columns to
remove. Uses the select() function from the dplyr package.
}
\examples{
\dontrun{# Keep columns from colData data frame
...  \%>\%
 mt_anno_select(anno_type="samples", term= ) \%>\%
 # remove columns from rowData data frame
 mt_anno_select(anno_type="features", term = )...}

}
\author{
JK, KC
}
