% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_xls.R
\name{mt_anno_xls}
\alias{mt_anno_xls}
\title{Load annotations from Excel file}
\usage{
mt_anno_xls(
  D,
  file,
  sheet,
  anno_type,
  anno_id_col,
  data_id_col = anno_id_col,
  no_map_err = F,
  replace_names_col = NA
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{file}{Name of input Excel file.}

\item{sheet}{Name or number of sheet.}

\item{anno_type}{Either "samples" (colData) or "features" (rowData).}

\item{anno_id_col}{Column in annotation file that contains ID information for mapping.}

\item{data_id_col}{Column in existing data that contains ID information for mapping. Default: Equal to anno_id_col.}

\item{no_map_err}{Throw error (T) or warning (F) if something does not map. Default: F.}

\item{replace_names_col}{OPTIONAL. Name of column from new annotation file to use to overwrite colnames (i.e. sample names) of SE. Default: none.}
}
\value{
rowData or colData: New annotation columns added.
}
\description{
Loads annotations and merges them into current SummarizedExperiment.
Performs "left-joins", i.e. leaves the original SE unchanged and just adds information where it can be mapped.
Can load annotations for both features (rowData) and samples (colData).

If annotation fields are already existing, this function will fill up any NAs with the values from the new file.
Existing values are not overwritten.
}
\examples{
\dontrun{
  # Load data, two sheets with sample annotations, and one sheet with feature annotations from the same file
  D <-
  # load raw data
  mt_load_xls(file=file, sheet="data", samples_in_rows=T, id_col="SAMPLE_NAME") \%>\%
  # sample annotations from metabolomics run
  mt_anno_xls(file=file, sheet="sampleinfo", anno_type="samples", anno_id_col = "SAMPLE_NAME") \%>\%
  # sample annotations from clinical table
  mt_anno_xls(file=file, sheet="clinicals", anno_type="samples", anno_id_col="SAMPLE_NAME") \%>\%
  # feature annotations`
  mt_anno_xls(file=file, sheet="metinfo", anno_type="features", anno_id_col="BIOCHEMICAL", data_id_col = "name") \%>\%
  ...}

}
\author{
JK
}
