% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_clean_fix_ggplot.R
\name{mt_clean_fix_ggplot}
\alias{mt_clean_fix_ggplot}
\title{Fix ggplot env objects}
\usage{
mt_clean_fix_ggplot(D)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}
}
\value{
$results[["plots"]]$output: Convert ggplot env object to image.
}
\description{
Convert all plots saved as large ggplot env objects to images.
}
\examples{
# example of how to run function
\dontrun{... \%>\%
  mt_clean_fix_ggplot() \%>\%
...}

}
\author{
KC
}
