% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_clean_remove_results.R
\name{mt_clean_remove_results}
\alias{mt_clean_remove_results}
\title{Remove metadata results from \code{SummarizedExperiment} object}
\usage{
mt_clean_remove_results(D, remove = "all")
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{remove}{What to remove. Has to be "all" or "plots". Default: "all".}
}
\value{
$results: Either all plots removed or all results entries removed.
}
\description{
Removes either the entire result list or all plots from the metadata of a \code{SummarizedExperiment}.
Can be used to obtain a more lightweight object for further processing.

remove=="all" will only leave a single result in the object, the one from this function.

remove=="plots" will set all ggplot objects to NULL, still allowing HTML reports to be generated.
}
\examples{
\dontrun{# at the end of a pipeline
... \%>\% mt_clean_remove_results(remove="plots")
...}

}
\author{
JK
}
