% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_clean_validate_se.R
\name{mt_clean_validate_se}
\alias{mt_clean_validate_se}
\title{Validate SE object is maplet-compatible}
\usage{
mt_clean_validate_se(D)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}
}
\value{
maplet-compatible SE
}
\description{
Validate an externally created SE object is maplet-compatible. If not, make compatible.
}
\examples{
# example of how to run function
\dontrun{... \%>\%
  mt_clean_validate_se() \%>\%
...}

}
\author{
KC
}
