% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_load_checksum.R
\name{mt_load_checksum}
\alias{mt_load_checksum}
\title{Validate MD5 checksum of file}
\usage{
mt_load_checksum(D, file, checksum)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input. Missing if first step in pipeline.}

\item{file}{File path. Does not have to be the same as the dataset loaded in the pipeline.}

\item{checksum}{Checksum to test for.}
}
\value{
If first step in pipeline, creates \code{SummarizedExperiment} object. Otherwise, does not change the
   \code{SummarizedExperiment} object.
}
\description{
If checksum provided, will check that the file checksum is the same and crash if not. If no checksum provided, stores the checksum
of the file in the pipeline log.
}
\examples{
# first call, to get the checksum (will crash, deliberately)
\dontrun{... \%>\% mt_load_checksum(file="input.xlsx", checksum="") \%>\% ...}

# copy-paste the correct ('actual') checksum from the error message into the call:
\dontrun{... \%>\% mt_load_checksum(file="input.xlsx", checksum="688048bd1eb9c771be0eb49548a6f947") \%>\% ...}

}
\author{
JK
}
