% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_load_merge_se.R
\name{mt_load_merge_se}
\alias{mt_load_merge_se}
\title{Merges Two maplet SE objects}
\usage{
mt_load_merge_se(
  D1,
  id_col1,
  D2,
  id_col2,
  crash_on_duplicate = T,
  sample_anno_from = "first"
)
}
\arguments{
\item{D1}{First \code{SummarizedExperiment} input to combine into; the one in the pipeline.}

\item{id_col1}{Name of the column in the colData of the first SE containing sample IDs.}

\item{D2}{Second \code{SummarizedExperiment} input to add.}

\item{id_col2}{Name of the column in the colData of the second SE containing sample IDs.}

\item{crash_on_duplicate}{Crash function if duplicate features are detected? If FALSE, suffices "_1" and "_2" will be added to
the end of the features in D1 and D2, respectively. Default: T.}

\item{sample_anno_from}{Which SE object to take sample annotations (colData) data frame from. Must be "first" or "second".
Default: "first".}
}
\value{
Merges second SE object with first SE object.

$result$output: metadata of second SE object.
}
\description{
Merges one maplet SE object provided as argument (D2) into another provided as input SE object (D1). Combines assay, rowData, 
and colData data frames and adds metadata of argument SE object (D2) into input SE object (D1). Will only work if sample IDs for
each SE are identical.
}
\examples{
\dontrun{D1 <-
  # merge D2 into D1
  mt_load_merge_se(id_col1="Sample_ID", D2=D2, id_col2="Sample_ID") \%>\%
  ...}

}
\author{
KC
}
