% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_load_merge_se_new.R
\name{mt_load_merge_se_new}
\alias{mt_load_merge_se_new}
\title{Merges Two maplet SE objects}
\usage{
mt_load_merge_se_new(
  D1,
  D2,
  samp_id_col1,
  samp_id_col2,
  feat_id_col1,
  feat_id_col2,
  samp_suffix1,
  samp_suffix2,
  feat_suffix1,
  feat_suffix2,
  anno_cols_identical = TRUE,
  results_from = c("none", "first", "second")
)
}
\arguments{
\item{D1}{First \code{SummarizedExperiment} object to combine.}

\item{D2}{Second \code{SummarizedExperiment} object to combine.}

\item{samp_id_col1}{From the first SE, name of the column in colData containing sample IDs.}

\item{samp_id_col2}{From the second SE, name of the column in colData containing sample IDs.}

\item{feat_id_col1}{From the first SE, name of the column in rowData containing feature IDs.}

\item{feat_id_col2}{From the second SE, name of the column in rowData containing feature IDs.}

\item{samp_suffix1}{Suffix to add to sample names in D1 in order to ensure uniqueness.}

\item{samp_suffix2}{Suffix to add to sample names in D2 in order to ensure uniqueness.}

\item{feat_suffix1}{Suffix to add to feature names in D1 in order to ensure uniqueness.}

\item{feat_suffix2}{Suffix to add to feature names in D2 in order to ensure uniqueness.}

\item{anno_cols_identical}{Flag indicating whether to force annotation columns for both rowData
and colData to be identical. If TRUE, function will crash if annotation column names are not
identical. If FALSE, function will keep both common and unique columns between the two SE
objects and warn the user to ensure that this will not create any downstream problems. In
cases where one or more sample / features are identical, this flag will be ignored and all
columns will be combined and kept separately. Default: TRUE.}

\item{results_from}{Which SE object to take result list. Must be one of: "none", "first", or "second".
Default: "none" (new pipeline).}
}
\value{
Merges assay, rowData, and colData of two SE objects.

$result$output: metadata of first, second or both SE objects.
}
\description{
Merges one maplet SE object provided as argument (D1) with another provided as input SE object
(D2). Combines assay, rowData, and colData data frames. Creates new metadata result list or
can continue from one of the existing lists from either SE input (D1 or D2).
This function can handle cases of unique, identical or overlapping samples and features. It also
allows the user to ensure the uniqueness of the sample and feature IDs from one or both of the
SE objects by adding suffixes. In cases of overlapping samples and features, the merge function
will attempt to combine assay data frames unless any of the sample-feature pairs contains
non-missing values for both D1 and D2 assays.
We recommend the user ensure that the colData and rowData annotation column names are
identical between D1 and D2. If this is not the case, annotation columns from each SE object
will be suffixed with d1 or d2, depending on the SE object of origin, and combined into one
large annotation data frame. This may create issues of annotation redundancy and the user must
ensure this will not create downstream issues. In the case of overlapping samples or features,
annotation columns will be suffixed with d1 and d2 and combined into one data frame as described
above to avoid the potential of overwritting annotaiton values.
}
\examples{
\dontrun{D1 <-
  # merge D2 into D1
  mt_load_merge_se(id_col1="Sample_ID", D2=D2, id_col2="Sample_ID") \%>\%
  ...}

}
\author{
KC
}
