% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_load_metabolights.R
\name{mt_load_metabolights}
\alias{mt_load_metabolights}
\title{Loader for MetaboLights Studies}
\usage{
mt_load_metabolights(
  D,
  study_id,
  met_file,
  samp_file,
  samp_col = "Sample.Name",
  samp_suffix,
  cache = FALSE,
  cache_file
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input. OPTIONAL.}

\item{study_id}{Identifier for the study of the form "MTBLS#" (where # is a number with 1 or more digits).}

\item{met_file}{Metabolite table file (prefix "m_") to download. Separates into assay and rowData data frames.
Use metabolighteR::get_study_files(study_id) for a list of files available.}

\item{samp_file}{Sample table file to download (prefix "s_"). Data for colData data frame. Use
metabolighteR::get_study_files(study_id) for a list of files available.}

\item{samp_col}{Name of column containing sample names in sample table. Default: "Sample.Name".}

\item{samp_suffix}{OPTIONAL. If sample names in metabolite table differ from names in samp_col by a suffix,
provide the suffix.}

\item{cache}{Cache downloaded data to local directory? Default: FALSE.}

\item{cache_file}{File for storing / retrieving cached data.}
}
\value{
If first step in pipeline, creates \code{SummarizedExperiment} object. Populates empty assay,
colData, and rowData data frames.
}
\description{
Load SummarizedExperiment data frames using the metabolighteR package to retrieve publicly available
data files from the metabolomics data repository Metabolights. If cache=TRUE, the assembled
SummarizedExperiment object (with variable name "D_cache") will be saved to a local directory. This loader
will always check for the presence of a cached file corresponding to the study ID and ion mode (optional)
provided. If found, this file will be used and new data will not be downloaded.
}
\examples{
\dontrun{
  # load data
  D1 <-mt_load_metabolights(study_id = "MTBLS301",
                            met_file = "m_MTBLS301_LC-MS__metabolite_profiling_v2_maf.tsv",
                            samp_file = "s_MTBLS301.txt")
  D2 <- mt_load_metabolights(study_id = "MTBLS264",
                             met_file = "m_mtbls264_POS_mass_spectrometry_v2_maf.tsv",
                             samp_file = "s_MTBLS264.txt",
                             samp_suffix="_POS")}

}
\author{
KC
}
