% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_load_metabolomics_workbench.R
\name{mt_load_metabolomics_workbench}
\alias{mt_load_metabolomics_workbench}
\title{Loader for Metabolomics Workbench Studies}
\usage{
mt_load_metabolomics_workbench(
  D,
  study_id,
  analysis_id,
  cache = FALSE,
  cache_file
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input. OPTIONAL.}

\item{study_id}{Identifier for the study of the form "ST######".}

\item{analysis_id}{Required if multiple datasets per study_id. Must be one of the form "AN######".}

\item{cache}{Use cached file to load data? Default: FALSE.}

\item{cache_file}{File for storing / retrieving cached data.}
}
\value{
If first step in pipeline, creates \code{SummarizedExperiment} object. Populates empty assay,
colData, and rowData data frames.
}
\description{
Load SummarizedExperiment data frames using the mwtab file provided by studies on the metabolomics
data repository Metabolomics Workbench.

Because of a formatting issue in the mwtab json objects, analysis IDs are required to distinguish
datasets in studies that provide multiple datasets.
}
\examples{
\dontrun{D <-
  # load data
  D1 <- mt_load_metabolomics_workbench(study_id = "ST000991")
  # use analysis_id to distinguish between multiple datasets
  D2 <- mt_load_metabolomics_workbench(study_id = "ST001301", analysis_id = "AN002167")
  }

}
\author{
KC
}
