% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_load_olink_v2.R
\name{mt_load_olink_v2}
\alias{mt_load_olink_v2}
\title{Load Olink-format data}
\usage{
mt_load_olink_v2(D, file)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input. Missing if first step in pipeline.}

\item{file}{Name of NPX file exported from NPX manger.}
}
\value{
Produces an initial \code{SummarizedExperiment}, with assay, colData, rowData, and metadata with first entry.
}
\description{
Loads data from an Olink-format Excel file.
Uses Olink's R code read_NPX taken from
\href{https://github.com/Olink-Proteomics/OlinkRPackage/tree/master/OlinkAnalyze}{https://github.com/Olink-Proteomics/Olink
RPackage/tree/master/OlinkAnalyze}.
}
\examples{
\dontrun{D <-
  # load data
  mt_load_olink(file=system.file("extdata", "example_data/sampledata.xlsx", package = "maplet")) \%>\%
  ...}

}
\author{
RB
}
