% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_load_se_xls.R
\name{mt_load_se_xls}
\alias{mt_load_se_xls}
\title{Load data matrix from Excel file - mt_write_se_xls format}
\usage{
mt_load_se_xls(D, file, sheet_names = c("assay", "rowData", "colData"))
}
\arguments{
\item{D}{\code{SummarizedExperiment} input. Missing if first step in pipeline.}

\item{file}{Name of input Excel file.}

\item{sheet_names}{Vector of the sheet names to read. Three sheet names must be provided and must
be ordered as such: (1) assay, (2) rowData, and (3) colData.
Default: c("assay", "rowData", "colData").}
}
\value{
If first step in pipeline, creates \code{SummarizedExperiment} object. Populates empty
        assay, colData, and rowData data frames.
}
\description{
Load data and annotation matrices from Excel sheet created by the mt_write_se_xls function. This
function is intended to be used on data produced by the mt_write_se_xls without manual
alteration.
}
\examples{
# Load data, rowData, and colData from Excel file
\dontrun{D <-
  # load raw data
  mt_load_se_xls(file=file, sheet_names=c("data", "metabolites", "samples")) \%>\%
  }

}
\author{
JK, KC
}
