% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_load_wcm.R
\name{mt_load_wcm}
\alias{mt_load_wcm}
\title{Load WCM core-format data}
\usage{
mt_load_wcm(D, file, sheet, zero_to_na = T)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input. Missing if first step in pipeline.}

\item{file}{Name of input Excel file.}

\item{sheet}{Name or number of sheet.}

\item{zero_to_na}{Replace zeros by NAs? Default: T.}
}
\value{
If first step in pipeline, creates \code{SummarizedExperiment} object. Populates empty assay, colData, and rowData data frames.
}
\description{
Loads data from a WCM core format file (as they send it).
}
\examples{
\dontrun{D <-
  # load data in WCM format
  mt_load_wcm(file='DM369_Metabolites separated by nucleotides 12_6_18_altered.xlsx',
              sheet='peakHeight_metabolite_intensiti') \%>\%
  ...}

}
\author{
JK
}
