% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_agg_pathways.R
\name{mt_modify_agg_pathways}
\alias{mt_modify_agg_pathways}
\title{Generate aggregated pathway values}
\usage{
mt_modify_agg_pathways(D, pw_col, method)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{pw_col}{Column name from rowData containing pathway annotations.}

\item{method}{One of: "eigen", "aggmean".}
}
\value{
assay: entirely replaces assay with new data matrix

rowData: removes original rowData, since now the "features" are pathways
}
\description{
Takes a pathway annotation column of the features (rowData) and builds one representative sample per pathway for each sample.
Also works for overlapping pathway annotations (i.e. where each feature can have >1 pathway).

Implemented approaches:
\enumerate{
  \item{Eigenmetabolite/eigengene/eigenvalue PCA based approach. Data matrix cannot have NAs.}
  \item{Mean value. Will NOT scale data before. Data matrix can have NAs.}
}
}
\examples{
\dontrun{\%>\% mt_modify_agg_pathways(pw_col="SUB_PATHWAY", method="aggmean") \%>\%  # subpathways from metabolon

# add KEGG pathways and use those
\%>\%
  mt_anno_pathways_hmdb(in_col = "HMDB", out_col = "kegg_db", pwdb_name = "KEGG", db_dir = system.file("extdata", "precalc/hmdb/", package = "maplet")) \%>\%
  mt_anno_pathways_remove_redundant(feat_id = "HMDB", pw_id = "kegg_db") \%>\%
  mt_modify_agg_pathways(pw_col="kegg_db", method="aggmean") \%>\%}

}
\author{
JK
}
