% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_avg_features.R
\name{mt_modify_avg_features}
\alias{mt_modify_avg_features}
\title{Averages duplicate features}
\usage{
mt_modify_avg_features(D, group_col)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{group_col}{Name of rowData column (feature annotation) by which duplicates can be identified.}
}
\value{
assay: Duplicate features (rows) combined.
}
\description{
Averages the values of features (rows) with the same values in specified rowData column.
}
\examples{
\dontrun{... \%>\% mt_modify_avg_features(group_col = 'name') \%>\% ...}

}
\author{
RB
}
