% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_avg_samples.R
\name{mt_modify_avg_samples}
\alias{mt_modify_avg_samples}
\title{Averages duplicate samples}
\usage{
mt_modify_avg_samples(D, group_col)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{group_col}{Name of colData column (sample annotation) by which duplicates can be identified.}
}
\value{
assay: Duplicate samples (columns) combined.
}
\description{
Averages the values of samples (columns) with the same values in specified colData column.
}
\examples{
\dontrun{... \%>\% mt_modify_avg_samples(group_col = "RID") \%>\% ...}

}
\author{
AS
}
