% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_diff_groups.R
\name{mt_modify_diff_groups}
\alias{mt_modify_diff_groups}
\title{Generate a new data matrix that is one group minus a second group (e.g. difference between two timepoints)}
\usage{
mt_modify_diff_groups(D, id_col, group_col, grp1, grp2)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{id_col}{Name of colData column representing sample ID variable  (e.g. patient ID).}

\item{group_col}{Name of colData column representing sample group variable (e.g. timepoint).}

\item{grp1}{Group level 1 (e.g. "day1").}

\item{grp2}{Group level 2 (e.g. "day2").}
}
\value{
assay: Replacement containing difference of the two groups.

colData: Replacement containing only those samples from grp1.
}
\description{
Takes the name of a sample factor variable (from colData), and two factor levels. Creates a new data matrix that
is made of the second minus the first levels. Will only produce results for cases where both data points were present in the
dataset.
}
\examples{
\dontrun{# Generate new dataset as differences between day2 and day1
... \%>\%  mt_modify_diff_groups(id_col = "patientID", group_col = "timepoint", grp1 = "day1", grp2 = "day2")  \%>\% ... # proceed with statistical analysis
}

}
\author{
JK
}
