% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_filter.R
\name{mt_modify_filter_samples}
\alias{mt_modify_filter_samples}
\title{Filter samples}
\usage{
mt_modify_filter_samples(D, filter)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{filter}{Logical expression. Can use columns from colData.}
}
\value{
assay: Filtered samples (columns) removed.

colData: Filtered samples removed.
}
\description{
Filters samples according to an expression. Expression can access entries of colData.
}
\examples{
# filter to two specific groups of samples
\dontrun{... \%>\% mt_modify_filter_samples(filter = GROUP \%in\% c("FL","ctrl")) \%>\% ...}

}
\author{
JK
}
