% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_ratios.R
\name{mt_modify_ratios}
\alias{mt_modify_ratios}
\title{Generate feature ratios}
\usage{
mt_modify_ratios(D, stat_name, edge_filter, neighborhood = 1)
}
\arguments{
\item{D}{\code{SummarizedExperiment} object.}

\item{stat_name}{Name of previous network generation call (e.g. \link{mt_stats_cormat_genenet}). Default: None (i.e. no
network-based ratios).}

\item{edge_filter}{Filter criterion for edge selection, e.g. "p.adj < 0.05", as a term.}

\item{neighborhood}{Neighborhood degree to use (e.g. first neighbors, second neighbors). Default: 1.}
}
\value{
assay: Pairwise ratios from all variables of input.

rowData: Pairwise ratios from all variables of input.
}
\description{
Transforms the dataset into a new dataset where each 'feature' represents a ratio of two features. Warning: For a dataset
with originally p features, this will result in p*(p-1) new variables (e.g. 500 features becomes 249500 ratios).

mt_post_pgain provides a special operation on a ratio data matrix for better interpretation of the resulting p-values.
}
\examples{
\dontrun{# Transform dataset to ratios
... \%>\%  mt_modify_ratios() \%>\% ... # proceed with statistical analysis
}

}
\author{
JZ, JK
}
