% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_box_scatter.R
\name{mt_plots_box_scatter}
\alias{mt_plots_box_scatter}
\title{Plot Box or Scatter Plots}
\usage{
mt_plots_box_scatter(
  D,
  x,
  stat_name,
  plot_type,
  correct_confounder,
  feat_filter = p.value < 0.05,
  feat_sort = p.value,
  annotation = "{sprintf('P-value: \%.1e', p.value)}",
  full_info = F,
  text_size = 3.88,
  jitter = "beeswarm",
  restrict_to_used_samples = T,
  ylabel = NULL,
  fit_line = T,
  fit_line_se = T,
  ggadd = NULL,
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{x}{Name of column from colData containing the phenotype to plot on x axis.}

\item{stat_name}{Index of the entry in metadata(D)$results that contains statistic object.}

\item{plot_type}{Either "box" or "scatter".}

\item{correct_confounder}{Confounders to adjust for before plotting, formula notation.}

\item{feat_filter}{If given, filter will be applied to data and remaining variables will be labeled in plot. Default: p.value<0.05.}

\item{feat_sort}{If given, arrange will be applied to data variables will be sorted. Default: p.value.}

\item{annotation}{If given adds annotation to plot. Default: p.value (see Usage for formatting).}

\item{full_info}{Add full information of all sample annotations and statistics results to plot table data.frame?
Makes plotting more flexible but can render SE objects huge. Default: F}

\item{text_size}{Text size of the annotations. Default: 3.88.}

\item{jitter}{Add geom_jitter ("jitter") or geom_beeswarm ("beeswarm") to boxplot. Exclude if NULL.  Default: "beeswarm".}

\item{restrict_to_used_samples}{Whether to filter to the samples that were used in the statistical test. Default: T.}

\item{ylabel}{Label for y-axis. Default: NULL.}

\item{fit_line}{Add fit line? Default: T.}

\item{fit_line_se}{Add standard error range? Default: T.}

\item{ggadd}{Further elements/functions to add (+) to the ggplot object. Default: NULL.}

\item{...}{Additional expression directly passed to aes() of ggplot, can refer to colData.}
}
\value{
if plot_type = "box", $result$output: plot, box plot

if plot_type = "scatter", $result$output: plot, scatter plot
}
\description{
Creates one box plot or scatter plot per feature based on given sample annotations.
}
\examples{
\dontrun{# boxplots as overview of results with a result already in 'comp'
# color by "Group" variable in colData
mt_plots_box_scatter(x                  = Group,
                     stat_name          = "comp",
                     plot_type          = "box",
                     correct_confounder = ~BATCH_MOCK,
                     feat_filter       = p.value<0.01,
                     feat_sort         = p.value,
                     annotation         = "{sprintf('P-value: \%.1e', p.value)}\nStatistic: {sprintf('\%.2f', statistic)}",
                    ) \%>\%
                 ...
 }

}
\author{
JZ, KC
}
