% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_box_special.R
\name{mt_plots_box_special}
\alias{mt_plots_box_special}
\title{Special Box Plots}
\usage{
mt_plots_box_special(
  D,
  stat_name,
  id_col,
  time_col,
  group_col,
  use_se = TRUE,
  feat_filter = p.value < 0.05,
  feat_sort = p.value,
  annotation = "{sprintf('P-value: \%.1e', p.value)}",
  text_size = 3.88,
  restrict_to_used_samples = T,
  ylabel = NULL,
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{stat_name}{Index of the entry in metadata(D)$results that contains statistic object.}

\item{id_col}{Name of colData column containing sample IDs. Value for mapping sample pairs.}

\item{time_col}{Name of colData column containing factors or ordinal numbers representing time points.}

\item{group_col}{OPTIONAL. Name of colData column for grouping samples. Value on x-axis.}

\item{use_se}{For three or more time-points, use standard error (TRUE) or standard deviation (FALSE)? Default: TRUE.}

\item{feat_filter}{If given, filter will be applied to data and remaining variables will be labeled in plot. Default: p.value<0.05.}

\item{feat_sort}{If given, arrange will be applied to data variables will be sorted. Default: p.value.}

\item{annotation}{If given adds annotation to plot. Default: p.value (see Usage for formatting).}

\item{text_size}{Text size of the annotations. Default: 3.88.}

\item{restrict_to_used_samples}{Whether to filter to the samples that were used in the statistical test. Default: T.}

\item{ylabel}{Label for y-axis. Default: NULL.}

\item{...}{Additional expression directly passed to aes() of ggplot, can refer to colData.}
}
\value{
$results$output: plot, special box plot
}
\description{
Create specialized box plots for paired samples (requires values for id_col and time_col) or time-resolved plots (requires values
for id_col, time_col, and group_col).
}
\examples{
\dontrun{mt_plots_repeating_test(stat_name = "comp2",
time_col           = Group,
group_col          = Sex,
id_col             = Individual,
fill               = Group,
feat_filter        = p.adj<0.2,
feat_sort          = fc)}

}
\author{
KC
}
