% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_dilution_factor.R
\name{mt_plots_dilution_factor}
\alias{mt_plots_dilution_factor}
\title{Plot dilution factors}
\usage{
mt_plots_dilution_factor(D, in_col, boxplot = T, ggadd = NULL)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{in_col}{Name of colData column (sample annotation) to correlation dilution factors with.}

\item{boxplot}{Produce boxplot (TRUE), or beeswarm plot (FALSE). Only relevant for factor comparison. Default: T.}

\item{ggadd}{Further elements / functions to add (+) to the ggplot object. Default: NULL.}
}
\value{
$result$output: plot, comparison with dilution factors
}
\description{
Correlate variable with dilution factors from quotient normalization.

\itemize{
\item for factors: will produce boxplot
\item for quantitative variable: will produce scatter plot
\item requires exactly one previous quotient normalization call in the pipeline
}
}
\examples{
\dontrun{\%>\% mt_plots_dilution_factor(in_col="group") \%>\% # compare with 'group' sample annotation
}

}
\author{
JK
}
