% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_heatmap.R
\name{mt_plots_heatmap}
\alias{mt_plots_heatmap}
\title{Heatmap plot by pheatmap}
\usage{
mt_plots_heatmap(
  D,
  scale_data = F,
  sym_zero = F,
  fun = function(x) {     if (scale_data)          return(scale(x))     x },
  silent = TRUE,
  return_gg = T,
  annotation_col = NA,
  annotation_row = NA,
  gg_scale = 1,
  gg_ymin = 1 - gg_scale,
  gg_xmin = 1 - gg_scale,
  gg_xmax = gg_scale,
  gg_ymax = gg_scale,
  ggadd = NULL,
  color = (grDevices::colorRampPalette(rev(RColorBrewer::brewer.pal(n = 7, name =
    "RdYlBu"))))(100),
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} object.}

\item{scale_data}{Scaling the data. Default: T.}

\item{sym_zero}{Make color scale symmetric around 0? (should only be used for scaled data). Default: F.}

\item{fun}{Function to transform/scale \code{t(assay(D))}, ie \code{mat = fun(t(assay(D)))} will be plotted.}

\item{silent}{Don't draw the table? A pheatmap argument, maplet uses a different default. Default: T.}

\item{return_gg}{Should pheatmap object be converted to gg object. Default: T.}

\item{annotation_col}{Data frame that specifies the annotations shown on the top of the heatmap.}

\item{annotation_row}{Data frame that specifices the annotations shown on the left side of the heatmap.}

\item{gg_scale}{Scaling of plot to be converted to gg object.}

\item{gg_ymin}{Minimum coordinate for ggplot y-axis. Default: 1-gg_scale.}

\item{gg_xmin}{Minimum coordinate for ggplot x-axis. Default: 1-gg_scale.}

\item{gg_xmax}{Maximum coordinate for ggplot x-axis. Default: gg_scale.}

\item{gg_ymax}{Maximum coordinate for ggplot y-axis. Default: gg_scale.}

\item{ggadd}{Further elements/functions to add (+) to the ggplot object.}

\item{color}{Vector of colors used in heatmap.}

\item{\dots}{See \code{pheatmap::pheatmap} for pheatmap arguments.}
}
\value{
$results$output: plot, heatmap
}
\description{

}
\examples{
\dontrun{D \%>\%
mt_plots_heatmap(annotation_row = c("SUPER_PATHWAY", "PLATFORM", "RI"),
                 annotation_col = c("GROUP_DESC","BATCH_MOCK","gender"),
                 fun = function(x) scale(exp(scale(x))),
                 clustering_distance_cols =  "correlation",
                 clustering_distance_rows = "minkowski")
 }

}
\author{
MB
}
