% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_evalute_ml.R
\name{mt_plots_ml_evaluate}
\alias{mt_plots_ml_evaluate}
\title{Plot Evaluation Measures for Repeated Machine Learning Predictions}
\usage{
mt_plots_ml_evaluate(
  D,
  ml_name,
  plot_measures = c("spec", "sens", "ppv"),
  cutoff = 0.5,
  pos_class,
  plots_per_fold = TRUE
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{ml_name}{Name under which the machine learning results are stored, must be unique to all other machine 
learning results.}

\item{plot_measures}{A vector of measures to include; can be one or more of the following: spec, sens, f1, 
acc, and/or ppv. Default: c("spec", "sens", "ppv").}

\item{cutoff}{Threshold cutoff for prediction probability class assignment; function will select threshold 
closest to provided value. Default: 0.5.}

\item{pos_class}{Which label to set as positive class.}

\item{plots_per_fold}{Whether to generate varying threshold plots per fold. Default: TRUE.}
}
\value{
results$ouput: List of evaluation measure plots.
}
\description{
This function creates plots of evaluation measures for repeated machine learning predictions. The types of plots
created vary depending on the cross-validation / sampling method used.Below is a list of plot types generated 
for each method:
{ADD LIST}
}
\examples{
\dontrun{
  
}

}
\author{
KC
}
