% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_multstats_heatmap.R
\name{mt_plots_multstats_heatmap}
\alias{mt_plots_multstats_heatmap}
\title{Heatmap of Multiple Statistical Objects}
\usage{
mt_plots_multstats_heatmap(
  D,
  stat_list = NA,
  cutoff,
  feat_anno_col,
  signif_mark = "•",
  mark_size = 20,
  show_mark = T,
  color_signif = F,
  filter_signif = F,
  cluster_rows = F,
  cluster_cols = F,
  show_colnames = F,
  main = "Stat Results Overview"
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{stat_list}{List of stat names to plot. Default: NA (for all stat objects in D).}

\item{cutoff}{The p.adj value significance cutoff.}

\item{feat_anno_col}{The rowData() column name to use as column annotation.}

\item{signif_mark}{Marker used to indicate significant p-values. Default: "•".}

\item{mark_size}{Size of the marker. Default: 20.}

\item{show_mark}{Show significance marker? Default: T.}

\item{color_signif}{Boolean. Color only significant values? Default: F.}

\item{filter_signif}{Show only features significant in one or more results? Default: F.}

\item{cluster_rows}{Cluster by rows? Default: F.}

\item{cluster_cols}{Cluster by cols? Default: F.}

\item{show_colnames}{Should column (feature) labels be included in the plot? Default: F.}

\item{main}{Main title of pheatmap. Default: "Stat Results Overview".}
}
\value{
results$output: heatmap of statistical results
}
\description{
Generate a heatamp for a list of statistical results using \code{pheatmap::pheatmap}.
}
\examples{
\dontrun{D <- D \%>\% mt_plots_multstats_heatmap(feat_anno_col = "SUPER_PATHWAY",
                                               cluster_cols = T,
                                               cluster_rows = T)}

}
\author{
KC
}
