% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_net.R
\name{mt_plots_net}
\alias{mt_plots_net}
\title{Creates a network object}
\usage{
mt_plots_net(
  D,
  stat_name,
  cor_filter = p.value < 0.05,
  node_coloring,
  html_outfile,
  height = 500
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{stat_name}{Name of the test to take correlations from.}

\item{cor_filter}{Filter for correlation values to plot. Default: p.value < 0.05.}

\item{node_coloring}{Name of the test to use for node coloring.}

\item{html_outfile}{Name of visnetwork html file. If empty, no html saved.}

\item{height}{OPTIONAL. Size in pixel of the plotting window size. Default: 500.}
}
\value{
$result$output: network ggplot

$result$output2: visnetwork plot
}
\description{
Creates a network visualization object using the visNetwork package.
}
\examples{
\dontrun{#' # in the context of a SE pipeline
... \%>\% mt_plots_net(stat_name = "xxx") \%>\% ...    # standard call
... \%>\% mt_plots_net(stat_name = "xxx", cor_filter = p.adj < 0.5, node_coloring="Li's", html_outfile="Network.html", height=800) \%>\% ...    # filters only significant correlations and colors the nodes according to the results in the indicated test, saves visnetwork to file
}

}
\author{
EB
}
