% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_pathview.R
\name{mt_plots_pathview}
\alias{mt_plots_pathview}
\title{Creates KEGG pathways visualization}
\usage{
mt_plots_pathview(
  D,
  gene_id = NULL,
  gene_data = NULL,
  gene_idtype = "entrez",
  metab_id = NULL,
  cpd_data = NULL,
  cpd_idtype = "kegg",
  stat_name,
  metab_filter,
  color_scale,
  color_range,
  show_only_filtered = FALSE,
  low = list(gene = "green", cpd = "yellow"),
  mid = list(gene = "gray", cpd = "gray"),
  high = list(gene = "red", cpd = "blue"),
  pathway_id,
  n_pathways,
  path_database = "./Pathview_database",
  path_output = "./Pathview_output",
  same_layer = TRUE,
  out_suffix = "pathview",
  add_pwname_suffix = FALSE,
  db_path = system.file("extdata", "precalc/pathview/KeggPathways.Rds", package =
    "maplet"),
  limit = list(gene = 1, cpd = 1),
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{gene_id}{Name of the rowData column containing the gene identifiers.}

\item{gene_data}{Either vector (single sample) or a matrix-like data (multiple sample). Vector should be numeric with gene IDs
as names or it may also be character of gene IDs. Character vector is treated as discrete or count data. Matrix-like data
structure has genes as rows and samples as columns. Row names should be gene IDs. Here gene ID is a generic concepts,
including multiple types of gene, transcript and protein uniquely mappable to KEGG gene IDs. KEGG ortholog IDs are also
treated as gene IDs as to handle metagenomic data. Check details for mappable ID types. Default gene_data=NULL.}

\item{gene_idtype}{The ID type used for the gene_data, case insensitive. Default gene_idtype="entrez", i.e. Entrez Gene,
which are the primary KEGG gene ID for many common model organisms. For other species, gene_idtype should be set to "KEGG"
as KEGG use other types of gene IDs. For the common model organisms (to check the list, do: data(bods); bods), you may also
specify other types of valid IDs. To check the ID list, do: data(gene_idtype.list); gene_idtype.list.}

\item{metab_id}{Name of the rowData column containing the feature identifiers.}

\item{cpd_data}{Same as gene_data, excpet named with IDs mappable to KEGG compound IDs. Over 20 types of IDs included in
CHEMBL database can be used here. Check details for mappable ID types. Default cpd_data=NULL. Note that gene_data and cpd_data
can't be NULL simultaneously.}

\item{cpd_idtype}{The ID type used for the cpd_data. Currently only works with "kegg".}

\item{stat_name}{Name of the statistics object to apply metab_filter to.}

\item{metab_filter}{If given, filter will be applied to data and only variables satisfying the condition will be included.}

\item{color_scale}{If given, this will be used to map colors to a continuous scale.}

\item{color_range}{A positive numeric value. If given, indicates the color range (-color_range, +color_range). If missing,
color_range will be determined internally.}

\item{show_only_filtered}{If TRUE generate pathway list only based on filtered variables, otherwise pathways will be generated
based on all variables.}

\item{low, mid, high}{Each is a list of two colors with "gene" and "cpd" as the names. This argument specifies the color spectra
to code gene_data and cpd_data. Default spectra (low-mid-high) "green"-"gray"-"red" and "yellow"-"gray"-"blue" are used for
gene_data and cpd_data respectively. The values for 'low, mid, high' can be given as color names ('red'), plot color index
(2=red), and HTML-style RGB, ("\#FF0000"=red).}

\item{pathway_id}{A character vector. The KEGG pathway ID(s), usually 5 digit, may also include the 3 letter KEGG species code.
If missing, the function will find all KEGG pathway annotations for the given KEGG identifiers.}

\item{n_pathways}{OPTIONAL. Number of pathways to output. Most populated pathway will be plotted first.}

\item{path_database}{The directory path of KEGG pathway data file (.xml) and image file (.png). If the path does not exist,
the function will create it. Default: "./Pathview_database" (subfolder in the current working directory).}

\item{path_output}{The directory path of the function output files. If the path does not exist, the function will create it.
Default: "./Pathview_output" (subfolder in the current working directory).}

\item{same_layer}{Controls if node colors are to be plotted in the same layer as the pathway graph. If FALSE, output
generation will be faster, but output plots will be larger in size.}

\item{out_suffix}{The suffix to be added after the pathway name as part of the output graph file. Default: "pathview".}

\item{add_pwname_suffix}{If TRUE will add the pathway name to the output filename. If FALSE, will use what stored
in out_suffix for all files. Default: FALSE.}

\item{db_path}{path to the pathway database to read annotations from.}

\item{limit}{A list of two numeric elements with "gene" and "cpd" as the names. This argument specifies the limit values for
gene.data and cpd.data when converting them to pseudo colors.}

\item{\dots}{See \code{pathview::pathview} for additional pathview arguments.}
}
\value{
pathview images (external)
}
\description{
Wrapper function for the \code{pathview::pathview} function. Maps and renders user data on relevant pathway graphs.
}
\examples{
\dontrun{# plot all pathways with at least one significant feature from the statistical comparison "comp" in them
mt_plots_pathview(D = D,
                  metab_id="KEGG_mapped",
                  stat_name = "comp",
                  color_scale = -sign(fc)*log10(p.adj),
                  color_range = -log10(0.01),
                  metab_filter = p.adj < 0.05,
                  show_only_filtered = TRUE,
                  path_database = "./Pathview_database",
                  path_output = "./results/pathview",
                  same_layer = F,
                  add_pwname_suffix = T
                  ) \%>\%
                  ...}

}
\author{
EB
}
