% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_pca.R
\name{mt_plots_pca}
\alias{mt_plots_pca}
\title{2D PCA of samples}
\usage{
mt_plots_pca(
  D,
  title = "PCA",
  scale_data = F,
  pc1 = 1,
  pc2 = 2,
  data_type = "scores",
  label_col = "",
  text_repel = T,
  ellipse = NA,
  exp_var_plot = F,
  store_matrices = F,
  ggadd = NULL,
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{title}{Title of the plot. Default: "PCA".}

\item{scale_data}{Scale data before plotting? Scale to mean 0, std 1. Default: F.}

\item{pc1}{First PC to plot. Default: 1 (PC1).}

\item{pc2}{Second PC to plot. Default: 2 (PC2).}

\item{data_type}{Data of type 'scores' or 'loadings'. Default: 'scores'.}

\item{label_col}{Column to use for labels. Default: ''.}

\item{text_repel}{Try to avoid all text overlaps when labeling? Default:T.}

\item{ellipse}{Confidence interval for ellipse. Default: NA (no ellipse).}

\item{exp_var_plot}{Add explained variance plot? Default: F.}

\item{store_matrices}{Store scores and loadings matrices in result structure? Default: F.}

\item{ggadd}{Further elements/functions to add (+) to the ggplot object. Default: NULL.}

\item{...}{Additional expression directly passed to aes() of ggplot, can refer to colData.}
}
\value{
results$output: plot(s)

results$output2: scores and loadings matrix
}
\description{
Create PCA plot of samples. Can be colored by any variable in colData.
}
\examples{
\dontrun{## PCA on scale_data, color and shape by "Group" variable in colData
... $>$ mt_plots_pca(scale_data=T, color=Group, shape=Group, title="PCA - scaled data") \%>\% ...
## PCA scores plot on non-scaled data, with ellipse and extra explained variance plot, and two ggadds (white background and centering of title)
mt_plots_pca(title="PCA scores", data_type = 'scores', scale_data=F, pc1=1, pc2=2, ellipse=0.95, exp_var_plot=T, ggadd = theme_bw() + theme(plot.title=element_text(hjust=0.5)))
}

}
\author{
JK
}
