% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_pval_hist.R
\name{mt_plots_pval_hist}
\alias{mt_plots_pval_hist}
\title{Generate p-value histogram}
\usage{
mt_plots_pval_hist(D, stat_list = NULL)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{stat_list}{Name of the statistical results. If not given, will generate histogram for all. Default: NULL.}
}
\value{
$result$output: plot, p-value histogram
}
\description{
Generates a p-value histogram, either for a given list of statistical results or for all statistical results.
}
\examples{
\dontrun{... \%>\% mt_plots_pval_hist() \%>\% ...  # for all
... \%>\% mt_plots_pval_hist(stat_list='comp') \%>\% ...  # for one
}

}
\author{
JK
}
