% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_pval_qq.R
\name{mt_plots_pval_qq}
\alias{mt_plots_pval_qq}
\title{Generate p-value qq plot}
\usage{
mt_plots_pval_qq(D, stat_name)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{stat_name}{Name of the statistical result.}
}
\value{
$result$output: plot, p-value qq plot
}
\description{
QQ plot against uniform distribution.
}
\examples{
\dontrun{... \%>\% mt_plots_pval_qq(stat_name='comp') \%>\% ...  # for one
}

}
\author{
JK
}
