% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_sgi.R
\name{mt_plots_sgi}
\alias{mt_plots_sgi}
\title{SGI analysis plot}
\usage{
mt_plots_sgi(
  D,
  outcomes,
  distance_function = "euclidean",
  hclust_method = "ward.D2",
  min_size,
  padj_thresh = 0.05
)
}
\arguments{
\item{D}{A \code{SummarizedExperiment} object.}

\item{outcomes}{A vector of column names from colData. The outcomes to be tested by SGI.
See \code{\link[sgi]{sgi_init}} for details.}

\item{distance_function}{The distance method to be passed to the \code{\link[stats]{dist}}
function. Refer to link for details. Default: "euclidean".}

\item{hclust_method}{The clustering method to be passed to the \code{\link[stats]{hclust}}
function. Refer to link for details. Default: "ward.D2".}

\item{min_size}{Minimum size for a cluster to be considered. If no value provided, set to
5\% of sample size. See \code{\link[sgi]{sgi_init}} for details.}

\item{padj_thresh}{Adjusted p-value threshold to filtering out significant enrichments.}
}
\value{
$result$output: list containing ggplot object
}
\description{
Wrapper function for the \code{plot_overview} function from the \code{sgi} package. Performs a
simple SGI analysis and plots all parts in one combined overview plot. Refer to
\code{\link[sgi]{plot_overview}} for details.
}
\examples{
\donttest{
 D \%>\%
 mt_plots_sgi(outcomes = c("AGE", "SEX", "BMI")
              distance_function = "euclidean",
              hclust_method = "ward.D2",
              min_size = 18,
              padj_thresh = 0.05
             ) \%>\% ...
 }

}
\author{
KC
}
