% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_stats_pathway_bar.R
\name{mt_plots_stats_pathway_bar}
\alias{mt_plots_stats_pathway_bar}
\title{Creates a pathway comparison bar plot}
\usage{
mt_plots_stats_pathway_bar(
  D,
  stat_list,
  feat_filter = p.value < 1,
  group_col = "SUB_PATHWAY",
  color_col = NULL,
  y_scale = "fraction",
  sort_by_y = FALSE,
  assoc_sign_col,
  add_empty = FALSE,
  keep_unmapped = FALSE,
  outfile = NULL,
  ggadd = NULL,
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{stat_list}{List of names of the statistics objects to be used for filtering}

\item{feat_filter}{Filter will be applied to data and remaining variables will be used to create plot}

\item{group_col}{The rowData variable used to aggregate variables.}

\item{color_col}{OPTIONAL. A rowData variable used to color barplot. Default NULL.}

\item{y_scale}{Plot percentage or frequency of variables. Values c("fraction","count"). Default "fraction".}

\item{sort_by_y}{Sort pathways in plot according to y_scale. Default FALSE.}

\item{assoc_sign_col}{OPTIONAL. Parameter to discriminate between positive and negative associations. Needs to be the name
of a column in the statistical results indicated by stat_list.}

\item{add_empty}{BOOLEAN. If TRUE adds also empty pathways to the barplot.}

\item{keep_unmapped}{BOOLEAN. If TRUE keep NULL values (specified as "Unmapped"). Default: FALSE.}

\item{outfile}{OPTIONAL. Excel filename to save data to}

\item{ggadd}{Further elements/functions to add (+) to the ggplot object.}

\item{...}{Additional expression directly passed to aes() of ggplot, can refer to colData.}
}
\value{
$result$output: plot, barplot
}
\description{
Create a pathway comparison bar plot using a list of statistical results.
}
\examples{
\dontrun{# Barplot as overview of results with a result already in 'comp'
... \%>\%
mt_plots_stats_pathway_bar(stat_list     = "comp",
 feat_filter = p.adj < 0.05,
 group_col    = "SUB_PATHWAY",
 color_col      = "SUPER_PATHWAYdevtools::install(codes.makepath("MT/maplet"))",
 y_scale       = "count",
 sort_by_y         = TRUE) \%>\%
 ...}

}
\author{
Elisa Benedetti
}
