% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_umap.R
\name{mt_plots_umap}
\alias{mt_plots_umap}
\title{2D UMAP of samples}
\usage{
mt_plots_umap(
  D,
  title = "UMAP",
  scale_data = F,
  label_col = "",
  text_repel = T,
  store_matrices = F,
  n_neighbors = 15,
  ggadd = NULL,
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{title}{Title of the plot. Default: "UMAP".}

\item{scale_data}{Scale data before plotting? Mean 0, std 1. Default: F.}

\item{label_col}{Column to use as label. Default: ''.}

\item{text_repel}{Try to avoid all text overlaps when labeling? Default:T}

\item{store_matrices}{Store scores in result structure? Default: F}

\item{n_neighbors}{The size of the local neighborhood UMAP will look at when attempting to learn the meanifold structure
of the data. Default: 15.}

\item{ggadd}{Further elements/functions to add (+) to the ggplot object. Default: NULL.}

\item{...}{Additional expression directly passed to aes() of ggplot. Can refer to colData.}
}
\value{
result output: plot(s)

result output2: scores and loadings matrix
}
\description{
Create a UMAP plot of samples. Can be colored by any variable in colData.
}
\examples{
\dontrun{## UMAP on scale_data, color and shape by "Group" variable in colData
... $>$ mt_plots_umap(scale_data=T, color=Group, shape=Group, title="UMAP - scaled data") \%>\% ...
}

}
\author{
JK
}
