% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_volcano.R
\name{mt_plots_volcano}
\alias{mt_plots_volcano}
\title{Creates a volcano plot}
\usage{
mt_plots_volcano(
  D,
  x = fc,
  stat_name,
  feat_filter = p.value < 0.05,
  xlabel = gsub("~", "", as.character(x)),
  vline = NA,
  hline,
  ggadd = NULL,
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{x}{Name of column in statistics table containing values to be plotted on the x-axis. Default: fc.}

\item{stat_name}{Name of the statistics object to plot.}

\item{feat_filter}{If given, filter will be applied to data and remaining variables will be labeled in plot}

\item{xlabel}{Label for the x-axis. Default: gsub("~","",as.character(x)).}

\item{vline}{Where to draw vertical line (for fold-change), has to be single value. Default: NA.}

\item{hline}{Where to draw horizontal line (for p-values), has to be an expression such as 'p.adj < 0.1'.}

\item{ggadd}{Further elements/functions to add (+) to the ggplot object. Default: NULL.}

\item{...}{Additional expression directly passed to aes() of ggplot, can refer to colData.}
}
\value{
$result$output: plot, volcano
}
\description{
Create a volcano plot using data from a statistical result.
}
\examples{
\dontrun{# Volcano plot as overview of results with a result already in 'comp'
... \%>\%
mt_plots_volcano(stat_name     = "comp",
 feat_filter = p.adj < 0.1,
 color       = p.value < 0.05) \%>\%
 ...}

}
\author{
JZ, JK
}
