% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_post_fold_change.R
\name{mt_post_fold_change}
\alias{mt_post_fold_change}
\title{Compute fold-change}
\usage{
mt_post_fold_change(D, stat_name, fun = function(x) {
    mean(x, na.rm = T)
})
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{stat_name}{Name of the statistical comparison.}

\item{fun}{Function with respect to which fold changes are computed (e.g. difference of the log means or median). Default:
function(x){mean(x,na.rm=T)}.}
}
\value{
$results[[stat_name]]$output: fc column added to statistical table
}
\description{
Add feature fold-changes to statistical results table.
}
\examples{
\dontrun{# add fold-changes to the result table of the statistical comparison called "comparison1", after correcting for variable "age"
... \%>\%
 mt_post_fold_change(stat_name="comparison1") \%>\% ...}


}
\author{
JZ
}
