% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_post_multtest.R
\name{mt_post_multtest}
\alias{mt_post_multtest}
\title{Multiple testing correction}
\usage{
mt_post_multtest(D, stat_name, p_col = p.value, method = "bonferroni")
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{stat_name}{Name of the statistical comparison to adjust.}

\item{p_col}{Name of p-value column to adjust. Default: p.value.}

\item{method}{Which method to use for multiple testing. See \code{p.adjust} documentation. Default: "bonferroni".}
}
\value{
$results[[stat_name]]$output: p.adj column added to statistical table.
}
\description{
Adjust output of statistical tests for multiple testing.
}
\examples{
\dontrun{# correct the statistical comparison called "Li's" using Benjamini-Hochberg
... \%>\%
 mt_post_multtest(stat_name="Li's", method="BH") \%>\% ...
 }

}
\author{
JK, JZ
}
