% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_post_pgain.R
\name{mt_post_pgain}
\alias{mt_post_pgain}
\title{Compute p-gain from feature ratio test}
\usage{
mt_post_pgain(D, stat_name, p_col = p.value)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{stat_name}{Name of the statistical comparison.}

\item{p_col}{Name of p-value column to compute p-gain from. Default: p.value.}
}
\value{
$results[[stat_name]]$output: pgain column added to statistical table.
}
\description{
This is a specialized post-analysis function for ratio-based datasets (see mt_modify_ratios). A "p-gain" is defined
as the factor of p-value decrease of a ratio, compared to the better of the p-values of the two single features. It
quantifies the amount of signal gained by working with the ratio.

For example, if feature A's association p-value is p=1e-5, feature B has p=1e-7, and the ratio has p=1e-9,
then pgain=min(1e-5, 1e-7)/1e-9 = 1e2 = 100.

See also: \href{https://pubmed.ncbi.nlm.nih.gov/22672667/}{https://pubmed.ncbi.nlm.nih.gov/22672667/}.
}
\examples{
\dontrun{# add p-gains to the result table of the statistical comparison called "comparison1"
... \%>\%
 mt_post_pgain(stat_name="comparison1") \%>\% ...
 }

}
\author{
JZ
}
