% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_batch_combat.R
\name{mt_pre_batch_combat}
\alias{mt_pre_batch_combat}
\title{ComBat batch correction}
\usage{
mt_pre_batch_combat(D, batch_col)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{batch_col}{Sample annotation (colData) column name that contains batch assignment.}
}
\value{
assay: Batch-corrected version.
}
\description{
Performs batch correction via COMBAT method from \code{sva} package. Function will check that data is logged. If data is not logged,
this function will log the data, run ComBat, and exponentiate the data before returning it into the assay data frame.
}
\examples{
\dontrun{... \%>\% mt_pre_batch_combat(batch_col="BATCH") \%>\% ...
}

}
\author{
JK
}
