% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_batch_median.R
\name{mt_pre_batch_median}
\alias{mt_pre_batch_median}
\title{Median batch correction}
\usage{
mt_pre_batch_median(D, batch_col, ref_samples)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{batch_col}{Sample annotation (colData) column name that contains batch assignment.}

\item{ref_samples}{Expression to filter out reference samples to use from colData.}
}
\value{
assay: Batch-corrected version.
}
\description{
Same approach that Metabolon uses for runday correction. Works for both logged and non-logged data. Subtracts (logged) or
divides (non-logged) feature values by the median value per batch.
}
\examples{
\dontrun{... \%>\% mt_pre_batch_median(batch_col="BATCH") \%>\% ...
}

}
\author{
JK
}
