% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_confounding_correction.R
\name{mt_pre_confounding_correction}
\alias{mt_pre_confounding_correction}
\title{Confounding correction}
\usage{
mt_pre_confounding_correction(D, formula, strata = NULL, scale_data = F)
}
\arguments{
\item{D}{\code{SummarizedExperiment} object.}

\item{formula}{Formula including column names from sample annotation, i.e, ~batch+age.}

\item{strata}{Strata classes for stratified confounding correction, should be a column from colData(D). Default: NULL.}

\item{scale_data}{Should data be scaled after correction? Default: F.}
}
\value{
assay: Corrected data.

$results$output: The lm.fit pvalue per features to show confounding affect on them.
}
\description{
Each feature is corrected by lm with the confounding variables and residuals are returned in lieu of uncorrected
features expressions.
}
\examples{
 \dontrun{# not run
 #... \%>\% mt_pre_confounding_correction( formula = ~batch + age, strata = "RUN_DAY" )
 }

}
\author{
MB
}
