% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_confounding_correction_stepaic.R
\name{mt_pre_confounding_correction_stepaic}
\alias{mt_pre_confounding_correction_stepaic}
\title{Confounding correction using stepwise AIC}
\usage{
mt_pre_confounding_correction_stepaic(D, cols_to_correct, n_cores = 1)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{cols_to_correct}{Vector of column numbers from colData to correct for.}

\item{n_cores}{Number of cores to use in parallelization. Default: 1.}
}
\value{
assay: Corrected data.

$results$output: Returns data.frame with the feature, its covars and the fit pval and rsq values.
}
\description{
Each feature is corrected by lm with the confounding variables determined by stepwise aic and residuals are returned
in lieu of uncorrected features expressions.
}
\examples{
 \dontrun{#... \%>\% mt_pre_confounding_correction_stepaic(cols_to_correct = c(1, 4, 5), n_cores = 10)
 }


}
\author{
AS, RB
}
