% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_cv.R
\name{mt_pre_cv}
\alias{mt_pre_cv}
\title{Calculate coefficient of variation (CV)}
\usage{
mt_pre_cv(D, qc_samples, out_col, replicates = F, id_col = NULL)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{qc_samples}{Logical expression. Can use columns from colData.}

\item{out_col}{Name of new rowData column to store the cv values.}

\item{replicates}{OPTIONAL. Flag to indicate if the selected samples are replicates. Default: F.}

\item{id_col}{OPTIONAL. If replicates is T, it is the name of the column in colData containing sample IDs. Default: NULL.}
}
\value{
rowData: New annotation column added.
}
\description{
Use QC samples to calculate coefficient of variation (CV) and add values as new column to rowData.
}
\examples{
\dontrun{... \%>\% mt_pre_cv(qc_samples=="PQC", out_col = "PQC_cv") \%>\% ...}

}
\author{
AS, RB
}
