% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_icc.R
\name{mt_pre_icc}
\alias{mt_pre_icc}
\title{Calculate intraclass correlation (ICC)}
\usage{
mt_pre_icc(D, qc_samples, out_col, id_col, icc_lmer = F)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{qc_samples}{Logical expression. Can use fields from colData.}

\item{out_col}{Name of new rowData column to store the icc values.}

\item{id_col}{Name of the colData column containing sample IDs.}

\item{icc_lmer}{OPTIONAL. Flag to set lmer flag in ICC function. Default: F.}
}
\value{
rowData: New annotation column added.
}
\description{
Uses QC samples to calculate intraclass correlation (ICC) and add values as new column to rowData.
}
\examples{
\dontrun{... \%>\% mt_pre_icc(qc_samples= (RID\%in\%RID[duplicated(RID)]),
out_col = "Dup_icc", id_col='RID') \%>\% ...}

}
\author{
RB
}
