% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_impute_min.R
\name{mt_pre_impute_min}
\alias{mt_pre_impute_min}
\title{Impute using minimum value}
\usage{
mt_pre_impute_min(D, verbose = F)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{verbose}{Whether to output intermediate steps. Default: F.}
}
\value{
assay: imputed data.
}
\description{
Impute using minimum value of that feature across the samples.
}
\examples{
\dontrun{# in the context of a SE pipeline
... \%>\% mt_pre_impute_min() \%>\%
}

}
\author{
RB
}
