% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_norm_quot.R
\name{mt_pre_norm_quot}
\alias{mt_pre_norm_quot}
\title{Quotient normalization}
\usage{
mt_pre_norm_quot(
  D,
  vars = 1:dim(D)[1],
  na_err = F,
  ref_samples = NULL,
  feat_max = 1
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{vars}{Index vector of variables to be used. Default: 1:dim(D)[1] (all).}

\item{na_err}{T/F, throw error for NA's or just ignore? Default: F.}

\item{ref_samples}{Expression filtering reference samples from colData. Default: NULL.}

\item{feat_max}{Maximum fraction of missingness to select features to be used in the reference. Default: 1 (i.e. all features).}
}
\value{
assay: Quotient-normalized version.

$results$output: List of dilution factors.
}
\description{
Implementation according to Dieterle et al., 2006\cr
\href{https://www.ncbi.nlm.nih.gov/pubmed/16808434}{https://www.ncbi.nlm.nih.gov/pubmed/16808434}.
}
\examples{
\dontrun{#' # in the context of a SE pipeline
... \%>\% mt_pre_norm_quot() \%>\% ...    # standard call
... \%>\% mt_pre_norm_quot(ref_samples = GROUP=="ctrl") \%>\% ...    # use reference samples where 'GROUP' field in colData is 'ctrl'
... \%>\% mt_pre_norm_quot(feat_max = 0.2) \%>\% ...    # use only features with <= 20\% missing values to compute the reference used for normalization
}

}
\author{
JK
}
