% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_outlier_detection_leverage.R
\name{mt_pre_outlier_detection_leverage}
\alias{mt_pre_outlier_detection_leverage}
\title{Identifies sample outliers using leverage approach}
\usage{
mt_pre_outlier_detection_leverage(D, reduce_dim = F, thresh = 4)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{reduce_dim}{If TRUE performs PCA-based dimensionality reduction before outlier detection. Can be used to apply multivariate
outlier detection methods to low-rank datasets.}

\item{thresh}{Number of m/n units to use as threshold to define an outlier. Default: 4.}
}
\value{
colData: New columns including a binary vector and a numeric score vector.

$results$output: Returns the specific parameters used to determine outliers.
}
\description{
Multivariate approach that uses leverage to define outliers. A sample is defined an outlier if its leverage is larger than
\code{thresh} times m/n (where m is the number of features and n is the number of samples in the dataset).
}
\examples{
\dontrun{# identify multivariate outliers with a leverage >4m/n
... \%>\%
  mt_pre_outlier_detection_leverage(thresh=4) \%>\%
...}

}
\author{
EB, JK
}
