% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_outlier_detection_univariate.R
\name{mt_pre_outlier_detection_univariate}
\alias{mt_pre_outlier_detection_univariate}
\title{Identifies Sample Outliers Using Univariate Approach}
\usage{
mt_pre_outlier_detection_univariate(D, reduce_dim = F, thresh = 4, perc = 0.5)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{reduce_dim}{If TRUE performs PCA-based dimensionality reduction before outlier detection. Can be used to apply multivariate
outlier detection methods to low-rank datasets.}

\item{thresh}{Number of standard deviations to use as threshold to define an outlier. Default: 4.}

\item{perc}{Ratio of features that need to be outliers in order to consider the whole sample an outlier. Default: 0.5.}
}
\value{
colData: New columns including a binary vector and a numeric score vector.

$results$output: Returns the specific parameters used to determine outliers.
}
\description{
A feature value is defined as an outlier if it is more than \code{thresh} standard deviations from the mean. A sample
is defined an outlier if more than \code{perc} of its features are univariate outliers.
}
\examples{
\dontrun{# identify samples that have more than 50\% univariate outliers
... \%>\%
  mt_pre_outlier_detection_univariate(thresh=4, perc=0.5) \%>\%
...}

}
\author{
EB, JK
}
