% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_outlier_lof.R
\name{mt_pre_outlier_lof}
\alias{mt_pre_outlier_lof}
\title{Remove outliers using local outlier factor}
\usage{
mt_pre_outlier_lof(D, seq_k = c(5, seq(10, nrow(df) - 1, by = 10)))
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{seq_k}{Sequence of numbers of nearest neighbors to use. Default: c(5, seq(10, nrow(df), by=10)).}
}
\value{
assay: Outlier samples removed.
}
\description{
Identifies outliers using the Local Outlier Factor method implmented in the \code{bigutilsr} package. 
See the following publication for details: 
\href{https://www.dbs.ifi.lmu.de/Publikationen/Papers/LOF.pdf}{https://www.dbs.ifi.lmu.de/Publikationen/Papers/LOF.pdf}.
For additional details about this method, including how to choose the threshold, refer to the following:
\href{https://www.r-bloggers.com/2019/08/detecting-outlier-samples-in-pca/}{https://www.r-bloggers.com/2019/08/detecting-outlier-samples-in-pca/}.
}
\examples{
\dontrun{
  D <- D \%>\% mt_pre_outlier_lof() \%>\%
...} 

}
