% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_outlier_to_na.R
\name{mt_pre_outlier_to_na}
\alias{mt_pre_outlier_to_na}
\title{Sets values of outliers in samples to NA}
\usage{
mt_pre_outlier_to_na(
  D,
  thresh = NA,
  sample_num_correction = T,
  alpha = 0.05,
  use_quant = FALSE,
  quant_thresh = 0.025,
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{thresh}{Number of standard deviations or m/n units to use as threshold to define the outlier, if not provided, default is
correction by sample size. This parameter is only used if sample_num_correction is False.}

\item{sample_num_correction}{Whether number of outliers should depend on number of samples or hard sd cutoff. If true, thresh is ignored}

\item{alpha}{The percentage of points we will consider outliers based on the assumption of a normal distribution (alpha/2 on either tail). Only used if sample_num_correction is True}

\item{use_quant}{Optional, if to flag the outliers in the quantile provided in the quantile parameter}

\item{quant_thresh}{If use_quant is true provide a quantile}
}
\value{
assay: NA values where outliers used to be
}
\description{
Uses univariate approach to identify outliers in the samples and sets the value of these outliers to NA.
}
\examples{
\dontrun{... \%>\%
  mt_pre_outlier_to_na(thresh=3, sample_num_correction=F) \%>\%
...}

}
\author{
AS, RB
}
