% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_reference_correction.R
\name{mt_pre_reference_correction}
\alias{mt_pre_reference_correction}
\title{Correct by reference}
\usage{
mt_pre_reference_correction(D, qc_samples, plate_col)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{qc_samples}{Logical expression. Can use fields from \code{colData()}}

\item{plate_col}{name of the colData column representing plate number}
}
\value{
D QCed
}
\description{
Correct data by a reference (e.g. NIST).
}
\examples{
\dontrun{... \%>\% mt_pre_reference_correction(qc_samples=Sample.Identification==102,
plate_col='Plate.Bar.Code') \%>\% ...}

}
\author{
RB
}
