% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_trans_exp.R
\name{mt_pre_trans_exp}
\alias{mt_pre_trans_exp}
\title{Exponentiate, base 2 by default}
\usage{
mt_pre_trans_exp(D, base = 2)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{base}{Operation: base^x for every data point. Default: 2.}
}
\value{
assay: Exponentiated data.
}
\description{
Transform the the entire dataset base^x.
}
\examples{
\dontrun{# in the context of a SE pipeline
... \%>\% mt_pre_trans_exp() \%>\% ...    # standard call, base 2
... \%>\% mt_pre_trans_exp(base=10) \%>\% ...    # base 10
}

}
\author{
JK
}
