% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_trans_log.R
\name{mt_pre_trans_log}
\alias{mt_pre_trans_log}
\title{Log, base 2 by default}
\usage{
mt_pre_trans_log(D, base = 2)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{base}{Operation: log(x)/log(base) for every data point. Default: 2.}
}
\value{
assay: Logged data.
}
\description{
Transform the entire dataset log_based(x), i.e. log(x)/log(base).
}
\examples{
\dontrun{# in the context of a SE pipeline
... \%>\% mt_pre_trans_log() \%>\% ...    # standard call, base 2
... \%>\% mt_pre_trans_log(base=10) \%>\% ...    # base 10
}

}
\author{
JK
}
