% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_trans_relative.R
\name{mt_pre_trans_relative}
\alias{mt_pre_trans_relative}
\title{Scale each sample relative to the mean of a given group of samples}
\usage{
mt_pre_trans_relative(D, ref_samples, is_logged)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{ref_samples}{Expression filtering reference samples from colData.}

\item{is_logged}{Is the data logged?}
}
\value{
assay: relatively scaled data
}
\description{
Usually used to make all concentrations relative to a control group, for example.
If data are logged, will use a 'minus' operation, if data are non-logged, will divide.
Whether data are logged is determine by whether they have been logged inside the MT pipeline.
}
\examples{
\dontrun{# normalize to control group, data not logged
... \%>\% mt_pre_trans_relative(ref_samples = GROUP=="ctrl", is_logged=F) \%>\% ...
}

}
\author{
JK
}
