% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_zero_to_na.R
\name{mt_pre_zero_to_na}
\alias{mt_pre_zero_to_na}
\title{Set zeros in dataset to NA}
\usage{
mt_pre_zero_to_na(D)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}
}
\value{
assay: Zeros replaced by NA.
}
\description{
Every 0 value in the dataset is replaced by an NA. Used for platforms that represent missing/sub-LOD values as zeros.
}
\examples{
\dontrun{# in the context of a SE pipeline
... \%>\% mt_pre_zero_to_na() \%>\% ...
}

}
\author{
JK
}
