% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_reporting_data.R
\name{mt_reporting_data}
\alias{mt_reporting_data}
\title{Output information about dataset to log}
\usage{
mt_reporting_data(D)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}
}
\value{
Does not change the \code{SummarizedExperiment} object.
}
\description{
Leaves a log entry containing the number of samples, features and annotation fields at the current stage of the pipeline.
}
\examples{
\dontrun{...  \%>\%
  mt_reporting_data() \%>\% ...}

}
\author{
JK
}
