% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_reporting_heading.R
\name{mt_reporting_heading}
\alias{mt_reporting_heading}
\title{Store heading for html report}
\usage{
mt_reporting_heading(D, heading, lvl = 1)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input. Missing if first step in pipeline.}

\item{heading}{Heading text.}

\item{lvl}{Heading level. Can be used for nested outline structures. Default: 1.}
}
\value{
If first step in pipeline, creates \code{SummarizedExperiment} object.

$result$output: Stores info about heading.
}
\description{
Store heading text and level to be used when calling \code{mt_reporting_html}.
}
\examples{
\dontrun{... \%>\%
# add first and second level of heading
mt_reporting_heading("Preprocessing") \%>\%
mt_reporting_heading("Part 1", lvl=2) \%>\%
...}

}
\author{
JK
}
