% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_reporting_html.R
\name{mt_reporting_html}
\alias{mt_reporting_html}
\title{Generates markdown-based HTML output from \code{SummarizedExperiment}}
\usage{
mt_reporting_html(
  D,
  file,
  title = "RMD output",
  output_calls = F,
  number_sections = F,
  start_after = NA,
  use_plotly = F,
  keep_tmp = F
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{file}{Output HTML filename.}

\item{title}{Title of RMD document. Default: 'RMD output'.}

\item{output_calls}{Output detailed info on function calls? Default: F.}

\item{number_sections}{Number sections and sub-sections? Default: F.}

\item{start_after}{Tag name or UUID of pipeline step AFTER which to start. Default: NA (i.e. output entire pipeline).}

\item{use_plotly}{EXPERIMENTAL. Output interactive plotly plots? WARNING: Setting this argument to TRUE can significantly
increase the runtime of the function. If plotly plots are not needed for a report, we recommend keeping this argument set to
FALSE. Default: F.}

\item{keep_tmp}{Keep temporary files? Can be used to manually edit RMD afterwards. Default: F.}
}
\value{
Does not change the \code{SummarizedExperiment} object.
}
\description{
Generates a fully automated report version of an entire (linear) pipeline.

Will generate RMD markdown document that can be adapted and then knitted to HTML.
}
\author{
JK
}
