% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_reporting_stats.R
\name{mt_reporting_stats}
\alias{mt_reporting_stats}
\title{Output information about statistical results}
\usage{
mt_reporting_stats(D, stat_name, stat_filter)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{stat_name}{Name of the statistical comparison.}

\item{stat_filter}{Filter formula to display number of results, e.g. p.adj<0.2.}
}
\value{
Does not change the \code{SummarizedExperiment} object.
}
\description{
Leaves a log entry containing the number of samples in each group that was used, and the number of hits according to a formula.
}
\examples{
\dontrun{...  \%>\%
  mt_reporting_stats(stat_name="comp1", stat_filter=p.adj<0.2) \%>\% ...}

}
\author{
JK
}
