% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_reporting_tag.R
\name{mt_reporting_tag}
\alias{mt_reporting_tag}
\title{Add a Reference Tag}
\usage{
mt_reporting_tag(D, tag_name)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{tag_name}{Name of the reference tag.}
}
\value{
$results$output: Add reference tag.
}
\description{
Add a reference tag to a section in the pipeline. For use by mt_reporting_html to indicate where in the pipeline the report should
start from.
}
\examples{
# add a tag to a section of the pipeline
\dontrun{
   # add tag to a spot in the pipeline
   mt_reporting_tag(tag_name="preprocessing") \%>\% ...
   D \%>\% mt_reporting_html(file="example.html", start_after="preprocessing")
}

}
\author{
KC
}
