% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_reporting_text.R
\name{mt_reporting_text}
\alias{mt_reporting_text}
\title{Store regular text for html report}
\usage{
mt_reporting_text(D, text)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{text}{Regular text.}
}
\value{
$results$output: Stores text.
}
\description{
Adds regular text to html report. Will be used when calling \code{mt_reporting_html}.
}
\examples{
\dontrun{... \%>\%
# add first and second level of heading
mt_reporting_heading("Preprocessing") \%>\%
mt_reporting_text("Data were first filtered for missing values, then normalized and log-transformed") \%>\%
...}

}
\author{
EB
}
