% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_reporting_tictoc.R
\name{mt_reporting_toc}
\alias{mt_reporting_toc}
\title{Timing functionality}
\usage{
mt_reporting_toc(D)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}
}
\value{
Does not change the \code{SummarizedExperiment} object.
}
\description{
Call mt_reporting_tic to start timing anywhere in pipeline, then mt_reporting_toc to show the time elapsed in the status log.
}
\examples{
\dontrun{\%>\%
  mt_reporting_tic() \%>\%
  ... \%>\% ... # pipeline steps
  ... \%>\% ... # pipeline steps
  ... \%>\% ... # pipeline steps
  mt_reporting_toc()}

}
\author{
JK
}
