% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_stats_cormat_genenet.R
\name{mt_stats_cormat_genenet}
\alias{mt_stats_cormat_genenet}
\title{Computes partial correlation matrix using the GeneNet estimator}
\usage{
mt_stats_cormat_genenet(D, stat_name, samp_filter)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{stat_name}{Name of the correlation matrix.}

\item{samp_filter}{Term defining which samples to use for GGM calculation. Default: missing (use all samples).}
}
\value{
$results$output: List of pairwise partial correlation coefficients and pvalues, as well as the corresponding variable names.
}
\description{
Implementation according to Schäfer and Strimmer, 2006\cr
\href{https://www.ncbi.nlm.nih.gov/pubmed/16646851}{https://www.ncbi.nlm.nih.gov/pubmed/16646851},
}
\examples{
\dontrun{... \%>\%
  mt_stats_cormat_genenet(stat_name ="pcor") \%>\%
...}

}
\author{
EB
}
