% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_stats_mixed_association_analysis.R
\name{mt_stats_mixed_association_analysis}
\alias{mt_stats_mixed_association_analysis}
\title{Mixed Association Analysis}
\usage{
mt_stats_mixed_association_analysis(
  D,
  stat_name,
  outcome,
  outcome_type,
  conf_formula = NULL
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{stat_name}{Name of the statistical result.}

\item{outcome}{Outcome to be used for response.}

\item{outcome_type}{Outcome type - one of numeric/binary/ordinal.}

\item{conf_formula}{Confounders to be corrected for. Default: NULL.}
}
\value{
$results$output: List of values including 'estimate', 'std_error', 'df', 'statistic', and 'p_value'.
}
\description{
ADD DESCRIPTION
}
\examples{
\dontrun{# ADD EXAMPLE}


}
\author{
RB
}
