% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_stats_pathway_enrichment.R
\name{mt_stats_pathway_enrichment}
\alias{mt_stats_pathway_enrichment}
\title{Pathway enrichment script using statistical analysis results from metabotools pipeline}
\usage{
mt_stats_pathway_enrichment(D, stat_name, pw_col, cutoff = 0.05)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{stat_name}{Name of statistical analysis to use with mti_get_stat_by_name().}

\item{pw_col}{Column containing pathways IDs.}

\item{cutoff}{Cutoff to use for assigning whether a metabolite/gene is significant. Used in Fisher's exact test. Default: 0.05.}
}
\value{
$results$pathways$enrichment_results: a dataframe containing the pathway enrichment results
}
\description{
This script runs a pathway enrichment analysis using Fisher's exact test from statistics
calculated by mt_stats_univ_lm().

Implemented approaches:
\enumerate{
   \item{Fisher's exact test.}
}
}
\examples{
\dontrun{\%>\% mt_stats_pathway_enrichment("kegg_db",
                                grouping_var = "Group",
                                control_grp_name = "Vehicle",
                                case_grp_name = c("treatment2", "treatment1") \%>\%
}


}
\author{
PG
}
