% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_stats_univ_cor.R
\name{mt_stats_univ_cor}
\alias{mt_stats_univ_cor}
\title{Computes correlation to a given phenotype}
\usage{
mt_stats_univ_cor(D, stat_name, in_col, method, samp_filter, exact = NULL)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{stat_name}{Name under which this comparison will be stored, must be unique to all other statistical results.}

\item{in_col}{Name of the colData variable to use for the correlation calculation. If method="kendall", class(D[[in_col]]) needs
to be numeric.}

\item{method}{Correlation method to use. Can be any among "pearson", "kendall", "spearman".}

\item{samp_filter}{OPTIONAL. Sample filter condition.}

\item{exact}{OPTIONAL. Set the exact flag in cor.test function. Default: NULL.}
}
\value{
$results$output: List of Kendall's correlation coefficients and pvalues, as well as the corresponding variable names.
}
\description{
Test for association between paired samples. If present, NAs will be omitted.
}
\examples{
\dontrun{... \%>\%
  mt_stats_univ_cor(in_col = "Stage", samp_filter = (GROUP \%in\% "Tumor"), stat_name = "tau", method = "tau") \%>\%
...
}

}
\author{
EB, RB
}
