% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_stats_univ_cs.R
\name{mt_stats_univ_cs}
\alias{mt_stats_univ_cs}
\title{Composite score models}
\usage{
mt_stats_univ_cs(D, in_col, id_col, stat_name, samp_filter)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{in_col}{Name of the column in colData to compare with features.}

\item{id_col}{Name of the column in colData having patient_ID.}

\item{stat_name}{Name under which this comparison will be stored, must be unique to all other statistical results.}

\item{samp_filter}{Term which samples to filter to first... e.g. used if the data contains >2 groups but the user wants to run a two-group comparison.}
}
\value{
$result$output: statistics object
}
\description{
Compute composite scores for outcome ~ features.
}
\examples{
\donttest{
 mt_stats_univ_cs(
   in_col     = Group,
   id_col = Subject_ID,
   stat_name         = "composite_score",
   samp_filter = (Group \%in\% c("grp1","grp2")),
 ) \%>\% ...
 }

}
\author{
RB
}
