% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_stats_univ_lm_matrixeqtl.R
\name{mt_stats_univ_lm_matrixeqtl}
\alias{mt_stats_univ_lm_matrixeqtl}
\title{Fast linear models using MatrixEQTL package}
\usage{
mt_stats_univ_lm_matrixeqtl(D, formula, stat_name, samp_filter)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{formula}{Right-hand side of formula to be put into glm function.}

\item{stat_name}{Name under which this comparison will be stored, must be unique to all other statistical results.}

\item{samp_filter}{Term which samples to filter to first (e.g. used if the data contains >2 groups but the user wants
to run a two-group comparison).}
}
\value{
$results$output: Statistics object.
}
\description{
Substantially faster than regular lm as implemented in mt_stats_univ_lm.
Only supports standard linear models of the form: outcome ~ feature + [covariates].
}
\examples{
\donttest{# run lm with no confounders, "Group" as outcome
# filter to groups "Li_2" and "Li_5"
# name the comparison "Li's"
... \%>\%
 mt_stats_univ_lm_matrixeqtl(
   formula      = ~ Group,
   samp_filter = (Group \%in\% c("Li_2","Li_5")),
   stat_name         = "Li's"
 ) \%>\% ...
 }

}
\author{
JK
}
