% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_stats_univ_missingness.R
\name{mt_stats_univ_missingness}
\alias{mt_stats_univ_missingness}
\title{Perform missingness significance analysis}
\usage{
mt_stats_univ_missingness(D, in_col, stat_name, samp_filter)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{in_col}{Sample annotation (colData) column to compare against.}

\item{stat_name}{Name of comparison for later reference.}

\item{samp_filter}{Sample filter term to restrict to.}
}
\value{
$results$output: Statistics object.
}
\description{
This function will determine if NAs significantly accumulate in one of the sample groups. It is recommended that this function
is run without prior missing value filtering.
}
\examples{
\dontrun{# run on sample field 'Group', name output stats object 'miss'
... \%>\% mt_stats_univ_missingness(in_col = 'Group', stat_name='miss') \%>\% ...
}

}
\author{
JK
}
