% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_stats_univ_wilcox.R
\name{mt_stats_univ_wilcox}
\alias{mt_stats_univ_wilcox}
\title{Wilcox test}
\usage{
mt_stats_univ_wilcox(
  D,
  in_col,
  stat_name,
  samp_filter,
  paired = F,
  exact = NULL
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{in_col}{Name of the column in colData to compare with feature.}

\item{stat_name}{Name under which this comparison will be stored, must be unique to all other statistical results.}

\item{samp_filter}{Term which samples to filter to first (e.g. used if the data contains >2 groups but the user
wants to run a two-group comparison).}

\item{paired}{OPTIONAL. Set the paired flag in wilcox.test function, would be applied for numeric in_col. Default: F.}

\item{exact}{OPTIONAL. Set the exact flag in wilcox.test function. Default: NULL.}
}
\value{
$results$output: Statistics object.
}
\description{
Perform Wilcoxon test.
}
\examples{
\donttest{# run lm with no confounders, "Group" as outcome
# filter to groups "Li_2" and "Li_5"
# name the comparison "Li's"
... \%>\%
 mt_stats_univ_wilcox(
   in_col     = Group,
   samp_filter = (Group \%in\% c("Li_2","Li_5")),
   stat_name         = "Li's"
 ) \%>\% ...
 }

}
\author{
RB
}
