% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_wrapper.R
\name{mt_wrapper}
\alias{mt_wrapper}
\title{Wrapper function for non-maplet code blacks}
\usage{
mt_wrapper(D, code)
}
\value{

}
\description{
This function allows the user to apply a block of non-maplet code on a maplet-SE object while
keeping the object within a maplet pipeline and ensuring the SE object is not altered in a way
that would make it maplet-incompatible. The code block must satisfy the following conditions:
\itemize{
   \item Must be passed as a function
   \item Function must take exactly one argument (corresponding to D)
   \item Function must not change dimensions of assay (this must be done with mt_modify_filter)
   \item Function must not filter the rows of colData or rowData (this must be done with
      mt_modify_filter); removal of columns is allowed
   \item Function must not change the colnames or rownames of the assay
   \item Function must return a list containing elements corresponding to these accepted names:
   \itemize{
      \item{df - assay}
      \item{cd - colData data frame}
      \item{rd - rowData data frame}
   }
}
}
\examples{
\dontrun{# remove unnecessary columns in colData
...  \%>\%
 mt_wrapper(code = my_code <- function(D){
    cd <- colData(D)
    cd <- dplyr::select(-c("GROUP_ID", "Age_at_diag", "C_Response", "Pathology_info"))
    list(cd = cd)
 }) \%>\% ...}

}
\author{
KC
}
