% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_write_pathways.R
\name{mt_write_pathways}
\alias{mt_write_pathways}
\title{Write out all pathway annotations, possible redundant pathways, and feature annotations}
\usage{
mt_write_pathways(D, pw_col, file)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{pw_col}{Name of the pathway annotation column.}

\item{file}{Output filename to write to.}
}
\value{
Does not change the \code{SummarizedExperiment} object.
}
\description{
Creates an Excel file that contains detailed feature-to-pathway mapping information. Writes out 4 different forms of the mapping
information. Also includes information about which pathways are redundant, i.e. contain the same set of features.
}
\examples{
\dontrun{... \%>\%
mt_anno_pathways_graphite(in_col = "KEGG",
                          out_col = "kegg_db",
                          pwdb_species = "hsapiens",
                          pwdb_name = "kegg") \%>\%
mt_anno_pathways_remove_redundant(feat_col = "KEGG", pw_col = "kegg_db") \%>\%
mt_write_pathways(pw_col='kegg_db', file='pwannos.xlsx')}

}
\author{
JK
}
