% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_write_se_rds.R
\name{mt_write_se_rds}
\alias{mt_write_se_rds}
\title{Save SummarizedExperiment to file}
\usage{
mt_write_se_rds(D, file)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{file}{Output filename to write to.}
}
\value{
Does not change the \code{SummarizedExperiment} object.
}
\description{
Simple helper function to save SE to a file. Object name will be \code{D}.
}
\examples{
\dontrun{#
... \%>\% mt_write_se_rds(file="out.RDS") \%>\% ...
...}

}
\author{
JK
}
