% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm_access.R
\name{mtm_get_assay_by_id}
\alias{mtm_get_assay_by_id}
\title{Return a specific assay version given a UUID or tag name}
\usage{
mtm_get_assay_by_id(D, uuid, tag_name)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{uuid}{A UUID associated with a specific function call.}

\item{tag_name}{Name of a tag given to a section of a maplet pipeline.}
}
\value{
An assay data frame.
}
\description{
Returns an assay version associated with a specific function call identified by the UUID or a tag name.
}
