% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm_access.R
\name{mtm_get_assay_by_ptr}
\alias{mtm_get_assay_by_ptr}
\title{Return a specific assay version given an assay pointer}
\usage{
mtm_get_assay_by_ptr(D, ptr)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{ptr}{A string used to identify assay versions.}
}
\value{
An assay data frame.
}
\description{
Returns an assay version identified by a specific assay pointer.
}
