% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm_access.R
\name{mtm_get_stat_by_name}
\alias{mtm_get_stat_by_name}
\title{Return stats output by name}
\usage{
mtm_get_stat_by_name(D, name, fullstruct = F)
}
\arguments{
\item{D}{SummarizedExperiment}

\item{name}{Name of statistical comparison}

\item{fullstruct}{optional, output entire $output structure, not just the $table inside.}
}
\value{
$output$table dataframe
}
\description{
Finds a named statistical result from a mt_stats... function and returns the $output$table dataframe.
}
