% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm_logging.R
\name{mtm_logging}
\alias{mtm_logging}
\title{Logging Function}
\usage{
mtm_logging(D = NULL, console = T)
}
\arguments{
\item{D}{define D here}

\item{console}{logging active or inactive?}
}
\value{
leaves SummarizedExperiment object unchanged
}
\description{
Put logging function description here
}
\author{
whoAuthoredIt?
}
