% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm_access.R
\name{mtm_res_get_entries}
\alias{mtm_res_get_entries}
\title{Extract results from entries in metadata by a list of name pieces}
\usage{
mtm_res_get_entries(D, name_list)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{name_list}{list of strings that represent pieces of function names}
}
\value{
List of results from the entries of interest.
}
\description{
Extracts metadata()$results entries in a given namespace of arbitrary depth. See examples below.
}
\examples{
\dontrun{
# get all result entries for plots
D \%>\% mtm_res_get_entries("plots")

# get all result entries for stats
D \%>\% mtm_res_get_entries("stats")

# check if quotient normalization has been performed
q <- D\%>\% mtm_res_get_entries(c("pre","norm","quot"))
if (length(q)==0) stop("No quotient normalization performed.")
}

}
\author{
JK
}
