% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm_access.R
\name{mtm_res_get_plots}
\alias{mtm_res_get_plots}
\title{Extract all plot objects from pipeline}
\usage{
mtm_res_get_plots(D, unlist = T)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{unlist}{Unlist all plots into one long list. Default: T.}
}
\value{
A list (nested or unlisted) containing all plots.
}
\description{
Returns just the plots, either retaining the results structure (a nested list) or unlisted.
}
