% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm_access.R
\name{mtm_res_get_plots_by_name}
\alias{mtm_res_get_plots_by_name}
\title{Extract plots from entries in metadata by a 'name argument'}
\usage{
mtm_res_get_plots_by_name(D, plot_name)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{plot_name}{Any name associated with a plot as a value of a function argument. 'Naming
arguments' include stat_name, stat_list, ml_name, stat1, stat2, and title.}
}
\value{
List of plots for the associated with the user-provided name.
}
\description{
Extracts plots from metadata()$results entries if they contain a 'naming argument' (see
description of plot_name argument) that is equal to the user provided name.
}
\examples{
\dontrun{
# make a volcano plot
D \%>\% mt_plots_volcano(stat_name = "Age super pw",
                       x = statistic,
                       feat_filter = p.adj < 0.05,
                       colour = p.adj < 0.05)

# get plot with 'name' "Age super pw"
D \%>\% mtm_res_get_plots_by_name("Age super pw")
}

}
\author{
KC
}
