% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm_access.R
\name{mtm_res_get_plots_by_type}
\alias{mtm_res_get_plots_by_type}
\title{Extract plots from entries in metadata by a type}
\usage{
mtm_res_get_plots_by_type(D, plot_type)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{plot_type}{A substring contained within the name of the function(s) of interest.}
}
\value{
List of plots for the plot types of interest.
}
\description{
Extracts plots from metadata()$results entries of functions that contain the user-provided
substring (via the plot_type argument) in their name. Note the substring can contain any portion
of the function name except the prefix 'mt_'. See examples.
}
\examples{
\dontrun{
# make a volcano and sample boxplots plot
D \%>\% mt_plots_volcano(stat_name = "Age super pw",
                        'x = statistic,
                        'feat_filter = p.adj < 0.05,
                        'colour = p.adj < 0.05) \%>\%
  mt_plots_sample_boxplot(color=Diagnosis, title = "Original", plot_logged = T) \%>\%
  mt_pre_batch_median(batch_col = "BOX.NUMBER") \%>\%
  mt_plots_sample_boxplot(color=Diagnosis, title = "After batch correction", plot_logged = T)

# get all volcano plots
D \%>\% mtm_res_get_plots_by_type("volcano")

# get all sample boxplots
D \%>\% mtm_res_get_plots_by_type("sample_boxplot")

# the following also returns all sample boxplots
D \%>\% mtm_res_get_plots_by_type("mple_boxpl")

# this doesn't work
D \%>\%  mtm_res_get_plots_by_type("mt_plots_sample_boxplot")
}

}
\author{
KC
}
