% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm_access.R
\name{mtm_res_get_ptrs}
\alias{mtm_res_get_ptrs}
\title{Extract all assay pointers for each function call}
\usage{
mtm_res_get_ptrs()
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}
}
\value{
A data frame with two columns: (1) call_order, (2) function_calls, (3) assay_ptrs.
}
\description{
Extracts all assay pointers for each function call and returns a data frame with each function call and corresponding 
assay pointer. The assay pointer is an identifier used to determine which assay version was used with a particular function
call. Will be NULL for all values if the global setting save_all_assays is FALSE.
}
