% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm_access.R
\name{mtm_res_get_uuids}
\alias{mtm_res_get_uuids}
\title{Extract all UUIDs for each function call}
\usage{
mtm_res_get_uuids(D)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}
}
\value{
A data frame with three columns: (1) call_order, (2) function_calls, (3) UUIDs.
}
\description{
Extracts all of the UUIDs for each function call and returns a data frame with each function call and corresponding UUID.
}
