% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt_compare_stepwise.R
\name{mtt_compare_stepwise}
\alias{mtt_compare_stepwise}
\title{Compare SE objects for equality for each step in a pipeline}
\usage{
mtt_compare_stepwise(file, ref_path, obj_pre)
}
\arguments{
\item{file}{Name of test script file.}

\item{ref_path}{Path to dir containing step-wise \code{SummarizedExperiment} input objects.}

\item{obj_pre}{Prefix of maplet pipeline test objects.}
}
\value{
Data frame with columns: step, fun, test, and result.
}
\description{
Compares reference objects for each step in a pipeline to the individual results of each step in a pipeline
being run in real-time.
Currently, plot objects and other non-stat table forms of output are ignored. Comparisons will only be made between two SE objects
with identical function calls.
}
\examples{
\dontrun{
     mtt_compare_stepwise(file = "stats_step_test.R", ref_path="REFERENCE_OBJECTS/", obj_pre = "stats_pipe")
}
}
