% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt_equal.R
\name{mtt_equal}
\alias{mtt_equal}
\title{Compare two SE objects for global equality}
\usage{
mtt_equal(D1, D2)
}
\arguments{
\item{D1}{First \code{SummarizedExperiment} input. When running tests, assume D1 is the reference SE object.}

\item{D2}{Second \code{SummarizedExperiment} input.}
}
\value{
Data frame with columns: test and result.
}
\description{
Compares the following attributes of two SE objects:
\itemize{
\item assay - checks assay data frames are identical
\item rowData - checks rowData data frames are identical
\item colData - checks colData data frames are identical
\item stat_len - checks both objects have the same number of statistical result tables
\item stat_# - checks statistical result tables are identical, # is a number identifying the statistical result
\item pathways_len - checks both objects have the same number of pathway annotation tables
\item pathways_# - checks pathway annotation data frames are identical, # is a number identifying the pathway annotation data frame
}
Currently, plot objects and other non-stat table forms of output are ignored. Comparisons will only be made between two SE objects 
with identical function calls.
}
\examples{
\dontrun{
     # compare two maplet pipleine objects
     mtt_equal(D1, D2)
}
}
