% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{D_ex2}
\alias{D_ex2}
\title{A toy SE object with duplicated metabolites}
\format{

}
\usage{
D_ex2
}
\description{
A \code{SummarizedExperiment} object with 50 samples and 50 metabolites. 5 out of the 100 samples are duplicated. The data for this object
is subset from the simulated_data.xlsx file.
}
\keyword{datasets}
