% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_load_nightingale.R
\name{mt_load_nightingale}
\alias{mt_load_nightingale}
\title{Load Nightingale-format data}
\usage{
mt_load_nightingale(
  D,
  file,
  data_sheet,
  format_type = "multiple_sheets",
  met_sheet = "Biomarker annotations",
  met_qc_sheet = "Tags per biomarker",
  samp_qc_sheet = "Quality control tags and notes",
  id_col
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input. Missing if first step in pipeline.}

\item{file}{Name of input excel file.}

\item{data_sheet}{If format_type is multiple sheets, name of sheet with data.}

\item{format_type}{Type of nightingale format: "single_sheet" or "multiple_sheets". Default: "multiple_sheets".}

\item{met_sheet}{If format type is multiple sheets, name of sheet with biomarker information. Default: "Biomarker annotations".}

\item{met_qc_sheet}{If format type is multiple sheets, name of sheet with biomarker qc tags. Default: "Tags per biomarker".}

\item{samp_qc_sheet}{If format type is multiple sheets, name of sheet with sample qc tags. Default: "Quality control tags and notes".}

\item{id_col}{OPTIONAL. Name of the column with sample IDs.}
}
\value{
Produces an initial \code{SummarizedExperiment}, with assay, colData, rowData, and metadata with first entry.
}
\description{
Loads data from a Nightingale format Excel file.
}
\examples{
\dontrun{D <-
  # load data
  mt_load_nightingale(file = =system.file("extdata", "example_data/sampledata.xlsx", package = "maplet")) \%>\%
  ...}

}
\author{
RB
}
