% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_load_olink.R
\name{mt_load_olink}
\alias{mt_load_olink}
\title{Load Olink-format data}
\usage{
mt_load_olink(D, file)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input. Missing if first step in pipeline.}

\item{file}{Name of Excel file exported from NPX manger.}
}
\value{
If first step in pipeline, creates \code{SummarizedExperiment} object. Populates empty assay (note: 2^NPX), colData, and rowData data frames.
}
\description{
Loads data from an Olink-format Excel file.
}
\examples{
\dontrun{D <-
  # load data
  mt_load_olink(file=system.file("extdata", "example_data/sampledata.xlsx", package = "maplet")) \%>\%
  ...}

}
\author{
RB, KC
}
