% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm_access.R
\name{mtm_res_get_plots}
\alias{mtm_res_get_plots}
\title{Extract all plot objects from pipeline}
\usage{
mtm_res_get_plots(D, unlist = T)
}
\arguments{
\item{D}{SummarizedExperiment}

\item{unlist}{Unlist all plots into one long list (default: T)}
}
\description{
Differs from mtm_res_get_plots_entries in that that function extracts full result structures, and this one returns just the plots
}
